MannKendall(openair)	R Documentation

Mann-Kendall trend tests

Description

Mann-Kendall tests for trend with Sen-Theil slope estimates.

Usage


MannKendall(mydata,
pollutant = "nox",
deseason = FALSE,
type = "default",
period = "monthly",
simulate = FALSE,
alpha = 0.05,
dec.place = 2,
ylab = pollutant,
main = "",
auto.text = TRUE,
autocor = FALSE, ...)

Arguments

mydata 	 A data frame containing the field date and at least one other parameter for which a trend test is required; typically (but not necessarily) a pollutant.

pollutant 	 The parameter for which a trend test is required. Mandatory.

deseason 	 Should the data be de-deasonalized first? If TRUE the function stl is used (seasonal trend decomposition using loess). Note that if TRUE missing data are first linearly interpolated because stl cannot handle missing data.

type 	 The type of analysis to be done. The default is will produce a single plot using the entire data. Other types include hour (for hour of the day), weekday (for day of the week), month (for month of the year), wd (wind direction split into eight sectors and ws (wind speed split into eight quantiles). There is an option site, which requires a slightly different format. If site is used a data frame with headings date, a pollutant and site is required. It is also possible to choose type as another variable in the data frame. For example, type = "o3" will plot four wind roses for different levels of ozone, split into four quantiles (approximately equal numbers of counts in each of the four splits). This offers great flexibility for understanding the variation of different variables dependent on another.

period 	 Either monthly (the default), or annual. Determines whether monthly mean or annual mean trends are plotted. Note that for annual, six or more years are required.

simulate 	 Should simulations be carried out to determine the Mann-Kendall tau and p-value. The default is FALSE. If TRUE, bootstrap simulations are undertaken, which also account for autocorrelation.

alpha 	 For the confidence interval calculations of the slope. The default is 0.05. To show 99% confidence intervals for the value of the trend, choose alpha = 0.01 etc.

dec.place 	 The number of decimal places to display the trend estimate at. The default is 2.

ylab 	 y-axis label.

main 	 Title of plot, if required.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the ‘2’ in NO2.

autocor 	 Should autocorrelation be considered in the trend uncertainty estimates? The default is FALSE. Generally, accounting for autocorrelation increases the uncertainty of the trend estimate - sometimes by a large amount.

... 	 Other graphical parameters.

Details

The Mann-Kendall function provides a collection of functions to analyse trends in air pollution data. The Mann-Kendall test is a commonly used test in environmental sciences to detect the presence of a trend. It is often used with the Sen-Theil (or just Sen) estimate of slope. See references. The Mann-Kendall function is flexible in the sense that it can be applied to data in many ways e.g. by day of the week, hour of day and wind direction. This flexibility makes it much easier to draw inferences from data e.g. why is there a strong downward trend in concentration from one wind sector and not another, or why trends on one day of the week or a certain time of day are unexpected.

The Mann-Kendall test from trend is for data that are monotonic - see http://en.wikipedia.org/wiki/Monotonic_function. The most appropriate use for this function is for data that are ‘well-behaved’ i.e. tend to be steadily increasing or decreasing or steady. For data that are strongly seasonal, perhaps from a background site, or a pollutant such as ozone, it will be important to deseasonalise the data (using the option deseason = TRUE.Similarly, for data that increase, then decrease, or show sharp changes it may be better to use smooth.trend.

Mann-Kendall code from Rand Wilxox http://www-rcf.usc.edu/~rwilcox/ has been modified to take account of correlated data.

The slope estimate and confidence intervals in the slope are plotted and numerical information presented.

The basic function have been adapted to take account of auto-correlated data using block bootstrap simulations (Kunsch, 1989). The principal reason for doing so is to gain a better estimate of trend uncertainty.

Warning

If deseason = TRUE, missing data will be linearly interpolated.

Author(s)

David Carslaw with trend code from Rand Wilcox

References

Helsel, D., Hirsch, R., 2002. Statistical methods in water resources. US Geological Survey. http://pubs.usgs.gov/twri/twri4a3/. Note that this is a very good resource statistics as applied to environmental data.

Hirsch, R. M., Slack, J. R., Smith, R. A., 1982. Techniques of trend analysis for monthly water-quality data. Water Resources Research 18 (1), 107-121.

Kunsch, H. R., 1989. The jackknife and the bootstrap for general stationary observations. Annals of Statistics 17 (3), 1217-1241.

... see also several of the Air Qulaity Expert Group (AQEG) reports for the use of similar tests applied to UK/European air quality data, see http://www.defra.gov.uk/ENVIRONMENT/airquality/panels/aqeg/.

See Also

See smooth.trend for a flexible approach to estimating trends using nonparametric regression. The smooth.trend function is suitable for cases where trends are not monotonic and is probably better for exploring the shape of trends.

Examples



# load example data from package
data(mydata)

# trend plot for nox
MannKendall(mydata, pollutant = "nox")

# trend plot for ozone with p=0.01 i.e. uncertainty in slope shown at
# 99 

MannKendall(mydata, pollutant = "o3", ylab = "o3 (ppb)", alpha = 0.01)

# trend plot by each of 8 wind sectors
## Not run: MannKendall(mydata, pollutant = "o3", type = "wd", ylab = "o3 (ppb)")

# and for a subset of data (from year 2000 onwards)
## Not run: MannKendall(select.by.date(mydata, year = 2000:2005), pollutant = "o3", ylab = "o3 (ppb)")


[Package openair version 0.1 Index]