Get information about airbase sites and instruments

Get information about airbase sites and instruments

airbaseInfo(code = "gb0620a", instrument = FALSE)

Arguments

code

Site code(s) of the sites to be imported. Can be upper or lower case.

instrument

Should species/instrument details also be returned. When FALSE only one row per site is returned with other information such as site type, latitude and logitude. When TRUE details of the individual species and instruments used is also returned.

Value

Returns a data frame containing key information about airbase site(s).

Details

This function compiles key information about airbase sites including the site type, species measured and the instruments and techniques used. By default the function returns one row per site with basic information concerning the site e.g. its type, latitude and longitude.

More comprehensive information is returned if instrument = TRUE. In this case the species measured at the site together with the instruments used will be returned.

See also

importAirbase, airbaseFindCode, airbaseStats