Bootsrap confidence intervals in the mean
Source: R/utilities.R

A utility function to calculation the uncertainty intervals in the mean of a vector. The function removes any missing data before the calculation.

bootMeanDF(x, conf.int = 0.95, B = 1000)

Arguments

x
A vector from which the mean and bootstrap confidence intervals in the mean are to be calculated

conf.int
The confidence interval; default = 0.95.

B
The number of bootstrap simulations

Value

Returns a data frame with the mean, lower uncertainty, upper uncertainty and number of values used in the calculation

Examples

test <- rnorm(20, mean = 10)
bootMeanDF(test)

#>          mean    min      max  n
#> Mean 9.897491 9.5324 10.25421 20