calc.fno2(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


calc.fno2(input, tau = 60, plot = TRUE, user.fno2, ...)

Arguments

input 	 ~~Describe input here~~

tau 	 ~~Describe tau here~~

plot 	 ~~Describe plot here~~

user.fno2 	 ~~Describe user.fno2 here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(input, tau = 60, plot = TRUE, user.fno2,...) {
{

library(mgcv)
library(lattice)
#function to prepare data ######################################################
prepare <- function(input) {

    #if cl or temp are missing, fill in with default values
    if(!any(names(input) %in% "temp"))  input$temp <- 11
    if(!any(names(input) %in% "cl"))  input$cl <- 4.5
    input$temp <- input$temp + 273

   #different date components
    input <- input[order(input$date), ]
    year <- as.numeric(format(input$date, "%Y"))
    month <- as.numeric(format(input$date, "%m"))
    jd <- as.numeric(format(input$date, "%j"))
    hour <- as.numeric(format(input$date, "%H"))

    nox.v <- input$nox - input$back_nox #estimated road contribution to NOx
    SLAT = 52 #latitute

    a <- 23.45 * (2 * pi / 360) * sin((jd + 284) * 2 * pi / 365)

    s <- sin(2 * pi * SLAT / 360) * sin(a) + cos(2 * pi * SLAT / 360) *
         cos(a) * cos((hour - 12) * 2 * pi / 24)

    solar <- (990 * s - 30) * (1 - 0.75 * (input$cl / 8) ^ 3.4)

    jno2 <- ifelse(solar < 0, 0, 0.0008 * exp(-10/solar) + 0.0000074*solar)

    k <-  0.04405*exp(-1370/input$temp)

    r <- jno2/k

    cbind(input, year, month, jd, hour, jno2, k, r, nox.v)
    
  }
################################################################################

calc.error <- function(user.fno2, nox, no2, nox.v, k, r, jno2, back_no2, back_o3, type) {
    
    #measured NOX-NO2 relationship
    bin1 <- cut(nox, breaks = 200)
    bin.meas <- aggregate(no2, list(bin1), mean, na.rm = TRUE)

    #first calculate new hourly means with t, f-NO2
    d <- 1/(k * tau)
    no2.v <-  nox.v * user.fno2
    no2.n <- no2.v + back_no2
    no2.o <- no2.n + back_o3
    b <- nox + no2.o + r + d
    no2.pred <- 0.5 * (b - (b^2 - 4 * (nox * no2.o + no2.n * d)) ^ 0.5)
    
    o3 <- (jno2 * tau * no2.pred + back_o3)/(k * tau * (nox - no2.pred) + 1)

    #make NOx-NO2 relationship
    bin.mod <- aggregate(no2.pred, list(bin1), mean, na.rm = TRUE)
    error <- sum(bin.meas$x - bin.mod$x, na.rm = TRUE)^2

    #return different results depending on use
    switch(type,
          err = error,
          conc = data.frame(no2 = no2.pred, o3 = o3)
    )
  }

#plot results ##################################################################
plot.fno2 <- function(results,...) {

      mod <- gam(fno2 ~ s(as.numeric(date)), data = results)
      pred <- predict(mod, results, se = TRUE)

      x1 <- c(results$date, rev(results$date))
      y1 <- c(pred$fit + 2 * pred$se.fit, rev(pred$fit - 2 * pred$se.fit))

      years <- as.numeric(unique(format(results$date, "%Y")))
      years <- c(years, max(years) + 1)
      years <- ISOdate(years, 1, 1)
      
      xyplot(fno2 ~ date, data = results,
          pch = 16,
          ylab = expression("f-NO" [2] * " (%)"),...,
          
          panel = function(x, y,...) {
              lpolygon(x1, y1, col = "lightblue", border = NA)
              panel.grid(h = -1, v = 0)
              panel.abline(v = as.Date(years), col = "grey85")
              panel.xyplot(x, y,...)
              panel.lines(results$date, pred$fit, col = "blue", lwd = 2,...)
          }
          )
      }

###plots orginal monthly NO2 and predicted with  ###############################
plot.no2 <- function(input, res,...) {

      results <- tapply(res$no2, format(res$date,"%Y-%m"), mean, na.rm = TRUE)
      measured <- tapply(input$no2, format(input$date,"%Y-%m"), mean, na.rm = TRUE)
      
      dates <- names(tapply(input$nox, format(input$date, "%Y-%m"), mean))
      dates <- as.Date(paste(dates, "-01", sep = ""))

      results <- data.frame(date = dates, no2 = results, no2.meas = measured)

      #modelled results
      mod <- gam(no2 ~ s(as.numeric(date)), data = results)
      pred.mod <- predict(mod, results, se = TRUE)

      x1 <- c(results$date, rev(results$date))
      y1 <- c(pred.mod$fit + 2 * pred.mod$se.fit, rev(pred.mod$fit - 2 * pred.mod$se.fit))
      
      #measured results
      meas <- gam(no2.meas ~ s(as.numeric(date)), data = results)
      pred.meas <- predict(meas, results, se = TRUE)

      x2 <- c(results$date, rev(results$date))
      y2 <- c(pred.meas$fit + 2 * pred.meas$se.fit, rev(pred.meas$fit - 2 * pred.meas$se.fit))

      years <- as.numeric(unique(format(results$date, "%Y")))
      years <- c(years, max(years) + 1)
      years <- ISOdate(years, 1, 1)

      xyplot(no2.meas ~ date, data = results,
          type = "l",
          ylab = expression("NO" [2] * " (ppb)"),

          key = list(lines = list(col = c("red", "blue"), lwd = 2),
                text = list(c("modelled", "measured")), columns = 2),...,

          panel = function(x, y,...) {
              lpolygon(x1, y1, col = rgb(1, 0, 0, 0.2), border = NA)
              lpolygon(x2, y2, col = rgb(0, 0, 1, 0.2), border = NA)
              panel.grid(h = -1, v = 0)
              panel.abline(v = as.Date(years), col = "grey85")
              panel.xyplot(x, y,...)
              panel.lines(results$date, pred.mod$fit, col = "red", lwd = 2)
              panel.lines(results$date, pred.meas$fit, col = "blue", lwd = 2)
          }
          )
  }
#start of code#################################################################
    input.all <- prepare(input)  #process input data
    input.all <- subset(input.all, nox.v > 0)  #process only +ve increments

    if(missing(user.fno2)) {

        fun.opt <- function(x)  {
                          optimize(calc.error, c(0, 0.5), x$nox, x$no2,
                                   x$nox.v, x$k, x$r, x$jno2, x$back_no2, x$back_o3, "err")
                   }
        #split data frame by year/month
        input.part <- split(input.all, input.all[c("month", "year")])
        
        input.part <- input.part[which(lapply(input.part, nrow) > 50)]  #need at least 50 hours
        
        fno2 <- sapply(input.part, function(x) fun.opt(x)$minimum)

        dates <- lapply(input.part, function(x) format(x$date[1], "%Y-%m"))
        dates <- do.call(rbind, dates)
        #dates <- names(tapply(input$nox, format(input$date, "%Y-%m"), mean))
        dates <- as.Date(paste(dates, "-01", sep = ""))

        results <- data.frame(date = dates, fno2 = 100 * fno2) # retrun estimates
        
        if (plot) print(plot.fno2(results,...))
        results
        
      } else {  ###calculate NO2 concentrations based on user input for whole series
         res <- calc.error(user.fno2, input.all$nox, input.all$no2,
                        input.all$nox.v, input.all$k, input.all$r,
                        input.all$jno2, input.all$back_no2,
                        input.all$back_o3, "conc")

        ids <- which(input$date %in% input.all$date)  #indices with data

        res <- cbind(date = input.all$date, res)

        gaps <- data.frame(date = input$date[-ids], no2 = input$no2[-ids], o3 = NA)

        res <- rbind(res, gaps)
        res <- res[order(res$date), ]
        if (plot) print(plot.no2(input.all, res,...))
        res
    }
  }
  }

[Package openair version 1.0 Index]