Import data from the UK Automatic Urban and Rural Network (AURN)

Function for importing hourly mean UK Automatic Urban and Rural Network (AURN) air quality archive data files for use with the openair package. Files are imported from a remote server operated by AEA that provides air quality data files as R data objects.

importAURN(site = "my1", year = 2009, pollutant = "all", hc = FALSE,
  verbose = TRUE)

Arguments

site

Site code of the AURN site to import e.g. "my1" is Marylebone Road. Several sites can be imported with site = c("my1", "nott") --- to import Marylebone Road and Nottingham for example.

year

Year or years to import. To import a sequence of years from 1990 to 2000 use year = 1990:2000. To import several specfic years use year = c(1990, 1995, 2000) for example.

pollutant

Pollutants to import. If omitted will import all pollutants ffrom a site. To import only NOx and NO2 for example use pollutant = c("nox", "no2").

hc

A few sites have hydrocarbon measurements available and setting hc = TRUE will ensure hydrocarbon data are imported. The default is however not to as most users will not be interested in using hydrocarbon data and the resulting data frames are considerably larger.

verbose

Should the function give messages when downloading files? Default is TRUE.

Value

Returns a data frame of hourly mean values with date in POSIXct class and time zone GMT.

Details

The importAURN function has been written to make it easy to import data from the UK AURN. AEA have provided .RData files (R workspaces) of all individual sites and years for the AURN. These files are updated on a daily basis. This approach requires a link to the Internet to work.

There are several advantages over the web portal approach where .csv files are downloaded. First, it is quick to select a range of sites, pollutants and periods (see examples below). Second, storing the data as .RData objects is very efficient as they are about four times smaller than .csv files --- which means the data downloads quickly and saves bandwidth. Third, the function completely avoids any need for data manipulation or setting time formats, time zones etc. Finally, it is easy to import many years of data beyond the current limit of about 64,000 lines. The final point makes it possible to download several long time series in one go. The function also has the advantage that the proper site name is imported and used in openair functions.

The site codes and pollutant names can be upper or lower case. The function will issue a warning when data less than six months old is downloaded, which may not be ratified.

The data are imported by stacking sites on top of one another and will have field names site, code (the site code) and pollutant. Sometimes it is useful to have columns of site data. This can be done using the reshape function --- see examples below.

All units are expressed in mass terms for gaseous species (ug/m3 for NO, NO2, NOx (as NO2), SO2 and hydrocarbons; and mg/m3 for CO). PM10 concentrations are provided in gravimetric units of ug/m3 or scaled to be comparable with these units. Over the years a variety of instruments have been used to measure particulate matter and the technical issues of measuring PM10 are complex. In recent years the measurements rely on FDMS (Filter Dynamics Measurement System), which is able to measure the volatile component of PM. In cases where the FDMS system is in use there will be a separate volatile component recorded as 'v10' and non-volatile component 'nv10', which is already included in the absolute PM10 measurement. Prior to the use of FDMS the measurements used TEOM (Tapered Element Oscillating. Microbalance) and these concentrations have been multiplied by 1.3 to provide an estimate of the total mass including the volatile fraction.

The few BAM (Beta-Attenuation Monitor) instruments that have been incorporated into the network throughout its history have been scaled by 1.3 if they have a heated inlet (to account for loss of volatile particles) and 0.83 if they do not have a heated inlet. The few TEOM instruments in the network after 2008 have been scaled using VCM (Volatile Correction Model) values to account for the loss of volatile particles. The object of all these scaling processes is to provide a reasonable degree of comparison between data sets and with the reference method and to produce a consistent data record over the operational period of the network, however there may be some discontinuity in the time series associated with instrument changes.

No corrections have been made to the PM2.5 data. The volatile component of FDMS PM2.5 (where available) is shown in the 'v2.5' column.

While the function is being developed, the following site codes should help with selection.

A3 	 London A3 Roadside 	 Urban traffic 	 -0.291853 	 51.37348 	 Kingston upon Thames London Boro 	 20/03/1997

30/09/2007 	 ABD 	 Aberdeen 	 Urban Background 	 -2.094278 	 57.15736 	 Aberdeen City

18/09/1999 	 30/09/2007 	 ABD7 	 Aberdeen Union Street Roadside 	 Urban traffic 	 -2.106472 	 57.14455

Aberdeenshire 	 01/01/2008 	 NA 	 ACTH 	 Auchencorth Moss 	 Rural Background 	 -3.242900

55.79216 	 Midlothian 	 04/09/2006 	 NA 	 AH 	 Aston Hill 	 Rural Background

-3.034178 	 52.50385 	 Powys - Powys 	 21/10/2003 	 NA 	 ARM6 	 Armagh Roadside

Urban traffic 	 -6.654528 	 54.35352 	 Armagh City and District Council 	 01/01/2009 	 NA 	 BALM

Ballymena 	 Urban Background 	 -6.258275 	 54.85864 	 Ballymena Borough Council 	 01/01/2010 	 NA

BAR2 	 Barnsley 12 	 Urban Background 	 -1.485153 	 53.55593 	 Barnsley District 	 21/03/1994

NA 	 BAR3 	 Barnsley Gawber 	 Urban Background 	 -1.510436 	 53.56292 	 Barnsley District

08/07/2002 	 30/09/2007 	 BARN 	 Barnsley 	 Urban Background 	 -1.473611 	 53.57972

Barnsley District 	 14/03/1991 	 21/03/1994 	 BATH 	 Bath Roadside 	 Urban traffic 	 -2.354155

51.39113 	 Bath and North East Somerset 	 18/11/1996 	 30/09/2007 	 BEL 	 Belfast East 	 Urban Background

-5.901667 	 54.59653 	 Belfast City Council 	 06/09/1989 	 30/09/2007 	 BEL2 	 Belfast Centre

Urban Background 	 -5.928833 	 54.59965 	 Belfast City Council 	 08/03/1992 	 NA 	 BEL4

Belfast Clara St 	 Suburban Background 	 -5.895460 	 54.59126 	 Belfast City Council 	 01/06/1998 	 30/09/2007

BEX 	 London Bexley 	 Suburban Background 	 0.184806 	 51.46603 	 London Borough of Bexley 	 01/05/1994

NA 	 BHAM 	 Birmingham Kerbside 	 Urban traffic 	 -1.906111 	 52.32778 	 Bromsgrove District

13/03/1974 	 17/01/1978 	 BIL 	 Billingham 	 Urban Industrial 	 -1.275039 	 54.60537

Stockton-on-Tees 	 01/01/1987 	 NA 	 BIR 	 Bircotes 	 Urban Background 	 -0.202500

51.88667 	 Bassetlaw District 	 14/09/1989 	 11/03/1991 	 BIR1 	 Birmingham Tyburn 	 Urban Background

-1.830583 	 52.51172 	 Birmingham District 	 23/01/2009 	 NA 	 BIR2 	 Birmingham East

Urban Background 	 -1.831498 	 52.49763 	 Birmingham District 	 04/08/1993 	 31/12/2000 	 BIRM

Birmingham Centre 	 Urban Background 	 -1.908078 	 52.47972 	 Birmingham District 	 18/03/1992 	 30/09/2007

BIRT 	 Birmingham Tyburn Roadside 	 Urban traffic 	 -1.830861 	 52.51219 	 Birmingham District 	 11/02/2009

NA 	 BLAC 	 Blackpool 	 Urban Background 	 -3.029283 	 53.79046 	 Blackpool

08/08/2000 	 11/11/2004 	 BLC2 	 Blackpool Marton 	 Urban Background 	 -3.007175 	 53.80489

Blackpool 	 05/06/2007 	 28/01/2009 	 BLCB 	 Blackburn Darwen Roadside 	 Urban traffic 	 -2.483815

53.71550 	 Blackburn with Darwen 	 15/06/2009 	 NA 	 BOLT 	 Bolton 	 Urban Background

-2.439583 	 53.57232 	 Bolton District 	 03/02/1997 	 30/09/2007 	 BORN 	 Bournemouth

Urban Background 	 -1.826744 	 50.73957 	 Bournemouth 	 19/07/2002 	 30/09/2007 	 BOT

Bottesford 	 Rural Background 	 -0.814722 	 52.93028 	 Melton District 	 01/10/1977 	 NA

BRAD 	 Bradford Centre 	 Urban Background 	 -1.748694 	 53.79339 	 Bradford District 	 28/11/1997

30/09/2007 	 BREN 	 London Brent 	 Urban Background 	 -0.276223 	 51.58977 	 London Borough of Brent

10/04/1996 	 30/09/2007 	 BRI 	 London Bridge Place 	 Urban Background 	 -0.141655 	 51.49521

City of Westminster 	 20/08/1990 	 30/11/1999 	 BRIS 	 Bristol Centre 	 Urban Background 	 -2.585622

51.45718 	 Bristol City Council 	 04/01/1993 	 15/09/2005 	 BRIT 	 Brighton Roadside 	 Urban traffic

-0.137281 	 50.82354 	 Brighton & Hove 	 10/02/1998 	 30/09/2007 	 BRN 	 Brentford Roadside

Urban traffic 	 -0.310121 	 51.48945 	 London Borough of Hounslow 	 20/06/2003 	 30/09/2007 	 BRS2

Bristol Old Market 	 Urban traffic 	 -2.583519 	 51.45603 	 Bristol City Council 	 01/07/1996 	 NA

BRS8 	 Bristol St Paul's 	 Urban Background 	 -2.584482 	 51.46284 	 Bristol City Council 	 15/06/2006

NA 	 BRT3 	 Brighton Preston Park 	 Urban Background 	 -0.147572 	 50.84084 	 Brighton & Hove

03/11/2004 	 NA 	 BURY 	 Bury Roadside 	 Urban traffic 	 -2.289611 	 53.53911

Bury District 	 20/01/1997 	 NA 	 BUSH 	 Bush Estate 	 Rural Background 	 -3.205782

55.86228 	 Midlothian 	 09/10/2003 	 NA 	 BY1 	 Bromley Roadside 	 Urban traffic

0.020128 	 51.40710 	 London Borough of Bromley 	 02/05/1997 	 06/07/1998 	 BY2 	 London Bromley

Urban traffic 	 0.075527 	 51.42468 	 London Borough of Bromley 	 11/08/1998 	 20/06/2005 	 CA1

Camden Kerbside 	 Urban traffic 	 -0.175269 	 51.54421 	 London Borough of Camden 	 19/02/2009 	 NA

CAM 	 Cambridge Roadside 	 Urban traffic 	 0.124456 	 52.20237 	 Cambridge District 	 26/06/1999

NA 	 CAMB 	 Cambridge 	 Urban traffic 	 0.041111 	 51.99528 	 North Hertfordshire District

01/11/1974 	 01/03/1978 	 CAN 	 London Canvey 	 Urban Industrial 	 0.571944 	 51.53250

Castle Point District 	 11/10/1977 	 25/02/1980 	 CANT 	 Canterbury 	 Urban Background 	 1.098061

51.27399 	 Canterbury District 	 01/02/2001 	 NA 	 CAR 	 Cardiff Kerbside 	 Urban traffic

-3.177778 	 51.48194 	 Cardiff County Council 	 28/07/1973 	 23/11/1976 	 CARD 	 Cardiff Centre

Urban Background 	 -3.176250 	 51.48178 	 Cardiff County Council 	 05/09/2002 	 10/09/2007 	 CARL

Carlisle Roadside 	 Urban traffic 	 -2.945861 	 54.89472 	 Carlisle City Council 	 17/03/2009 	 NA

CHAT 	 Chatham Roadside 	 Urban traffic 	 0.548668 	 51.37384 	 Medway 	 01/07/2010

NA 	 CHIL 	 Chilworth 	 Suburban Background 	 -1.422778 	 50.96222 	 Southampton City Council

18/04/1975 	 10/10/1975 	 CHP 	 Chepstow A48 	 Urban traffic 	 -2.678731 	 51.63809

Monmouthshire County 	 01/01/2008 	 NA 	 CHS6 	 Chesterfield 	 Urban Background 	 -1.433611

53.23058 	 Chesterfield District 	 13/03/2008 	 NA 	 CHS7 	 Chesterfield Roadside 	 Urban traffic

-1.456944 	 53.23172 	 Chesterfield District 	 11/03/2008 	 NA 	 CLL 	 Central London

Urban Background 	 -0.138333 	 51.49472 	 City of Westminster 	 01/07/1972 	 06/08/1990 	 CLL2

London Bloomsbury 	 Urban Background 	 -0.125889 	 51.52229 	 London Borough of Camden 	 23/01/1992 	 NA

COV2 	 Coventry Centre 	 Urban Background 	 -1.522133 	 52.41345 	 Coventry City Council 	 18/02/1997

31/12/2000 	 COV3 	 Coventry Memorial Park 	 Urban Background 	 -1.519612 	 52.39440 	 Coventry City Council

16/12/2008 	 NA 	 CRD 	 London Cromwell Road 	 Urban traffic 	 -0.180564 	 51.49492

Royal Borough of Kensington and Chelsea 	 22/02/1973 	 12/09/1996 	 CRD2 	 London Cromwell Road 2 	 Urban traffic 	 -0.178709

51.49548 	 Royal Borough of Kensington and Chelsea 	 20/05/1998 	 NA 	 CWMB 	 Cwmbran 	 Urban Background

-3.006953 	 51.65380 	 Tor-Faen - Torfaen 	 12/03/2002 	 30/09/2007 	 DERY 	 Derry

Urban Background 	 -7.329115 	 55.00122 	 Derry City Council 	 29/04/1997 	 30/09/2007 	 DUMB

Dumbarton Roadside 	 Urban traffic 	 -4.558333 	 55.94333 	 West Dunbartonshire 	 01/09/2010 	 NA

DUMF 	 Dumfries 	 Urban traffic 	 -3.614233 	 55.07003 	 Dumfries and Galloway 	 17/07/2001

30/09/2007 	 EAGL 	 Stockton-on-Tees Eaglescliffe 	 Urban traffic 	 -1.358547 	 54.51667 	 Stockton-on-Tees

21/01/2009 	 NA 	 EB 	 Eastbourne 	 Urban Background 	 0.271611 	 50.80578

Eastbourne District 	 01/11/2010 	 NA 	 ECCL 	 Salford Eccles 	 Urban Industrial 	 -2.334139

53.48481 	 Salford City Council 	 20/03/1997 	 NA 	 ED 	 Edinburgh Centre 	 Urban Background

-3.195775 	 55.95197 	 City of Edinburgh 	 04/10/1992 	 13/10/2003 	 ED3 	 Edinburgh St Leonards

Urban Background 	 -3.182186 	 55.94559 	 City of Edinburgh 	 24/11/2003 	 NA 	 EK

East Kilbride 	 Suburban Background 	 -4.173056 	 55.75472 	 South Lanarkshire 	 08/05/1975 	 09/10/1975

ESK 	 Eskdalemuir 	 Rural Background 	 -3.206111 	 55.31531 	 Dumfries and Galloway 	 09/12/2004

NA 	 EX 	 Exeter Roadside 	 Urban traffic 	 -3.532465 	 50.72508 	 Exeter District

02/07/1996 	 NA 	 FEA 	 Featherstone 	 Urban Background 	 -1.352500 	 53.68056

Wakefield District 	 14/09/1989 	 13/03/1991 	 FW 	 Fort William 	 Suburban Background 	 -5.101102

56.82266 	 West Lothian 	 22/06/2006 	 NA 	 GDF 	 Great Dun Fell 	 Rural Background

-2.450799 	 54.68423 	 Eden District 	 09/05/1986 	 NA 	 GLA 	 Glasgow City Chambers

Urban Background 	 -4.245959 	 55.86041 	 City of Glasgow 	 28/07/1989 	 30/09/2007 	 GLA3

Glasgow Centre 	 Urban Background 	 -4.255161 	 55.85773 	 City of Glasgow 	 26/07/1996 	 NA

GLA4 	 Glasgow Kerbside 	 Urban traffic 	 -4.258889 	 55.85917 	 City of Glasgow 	 01/08/2002

31/12/2010 	 GLAS 	 Glasgow Hope St 	 Urban traffic 	 -4.257778 	 55.85833 	 City of Glasgow

27/06/1973 	 28/10/1982 	 GLAZ 	 Glazebury 	 Rural Background 	 -2.472056 	 53.46008

Salford City Council 	 26/01/2004 	 NA 	 GRA2 	 Grangemouth Moray 	 Rural Background 	 -3.710833

56.01314 	 Falkirk 	 01/06/2009 	 NA 	 GRAN 	 Grangemouth 	 Urban Industrial

-3.704399 	 56.01032 	 Falkirk 	 17/01/2003 	 30/09/2007 	 HAR 	 Harwell

Rural Background 	 -1.325283 	 51.57108 	 Vale of White Horse District 	 14/08/1998 	 NA 	 HARR

London Harrow 	 Suburban Background 	 -0.348889 	 51.57333 	 London Borough of Harrow 	 24/08/1979 	 03/10/1980

HG1 	 Haringey Roadside 	 Urban traffic 	 -0.068225 	 51.59930 	 London Borough of Haringey 	 16/05/1996

NA 	 HG2 	 London Haringey 	 Urban Background 	 -0.126439 	 51.58605 	 London Borough of Haringey

29/11/2007 	 NA 	 HIL 	 London Hillingdon 	 Urban Background 	 -0.460861 	 51.49633

London Borough of Hillingdon 	 02/08/1996 	 30/09/2007 	 HK4 	 London Hackney 	 Urban Background 	 -0.056592

51.55877 	 London Borough of Hackney 	 06/01/1997 	 30/09/2007 	 HM 	 High Muffles 	 Rural Background

-0.808550 	 54.33494 	 Ryedale District 	 20/10/2003 	 NA 	 HOPE 	 Stanford-le-Hope Roadside

Urban traffic 	 0.439556 	 51.51817 	 Thurrock 	 22/01/2008 	 NA 	 HORE

Horley 	 Suburban Background 	 -0.167741 	 51.16586 	 Reigate and Banstead District 	 21/11/2007 	 NA

HORS 	 London Westminster 	 Urban Background 	 -0.131931 	 51.49467 	 City of Westminster 	 17/07/2001

NA 	 HOVE 	 Hove Roadside 	 Urban traffic 	 -0.170294 	 50.82778 	 Brighton & Hove

15/10/1997 	 30/09/2007 	 HR3 	 London Harrow Stanmore 	 Urban Background 	 -0.298777 	 51.61733

London Borough of Harrow 	 16/12/2008 	 NA 	 HRL 	 London Harlington 	 Urban Industrial 	 -0.441614

51.48879 	 London Borough of Hillingdon 	 01/01/2004 	 03/03/2008 	 HS1 	 Hounslow Roadside 	 Urban traffic

-0.308975 	 51.48965 	 London Borough of Hounslow 	 16/09/1997 	 16/11/2002 	 HUL2 	 Hull Freetown

Urban Background 	 -0.341222 	 53.74878 	 Kingston-upon-hull City Council 	 06/11/2002 	 NA 	 HULL

Hull Centre 	 Urban Background 	 -0.338322 	 53.74479 	 Kingston-upon-hull City Council 	 04/01/1994 	 17/01/2002

INV2 	 Inverness 	 Urban traffic 	 -4.241039 	 57.48120 	 Highland 	 17/07/2001

30/09/2007 	 ISL 	 London Islington 	 Urban Background 	 -0.093611 	 51.53083 	 London Borough of Islington

09/07/1976 	 11/10/1978 	 KC1 	 London N. Kensington 	 Urban Background 	 -0.213431 	 51.52106

Royal Borough of Kensington and Chelsea 	 31/03/2009 	 NA 	 LB 	 Ladybower 	 Rural Background 	 -1.752006

53.40337 	 High Peak District 	 27/10/1988 	 NA 	 LDS 	 Leeds Potternewton 	 Urban Background

-1.535098 	 53.82568 	 Leeds City Council 	 01/01/1995 	 31/12/2000 	 LEAM 	 Leamington Spa

Urban Background 	 -1.533119 	 52.28881 	 Warwick District 	 26/07/1996 	 30/09/2007 	 LED6

Leeds Headingley Kerbside 	 Urban traffic 	 -1.576361 	 53.81997 	 Leeds City Council 	 02/04/2009 	 NA

LEED 	 Leeds Centre 	 Urban Background 	 -1.546472 	 53.80378 	 Leeds City Council 	 04/01/1993

NA 	 LEIC 	 Leicester Centre 	 Urban Background 	 -1.133006 	 52.63135 	 Leicester City Council

04/01/1994 	 NA 	 LEOM 	 Leominster 	 Suburban Background 	 -2.736665 	 52.22174

Herefordshire Council 	 18/07/2005 	 NA 	 LERW 	 Lerwick 	 Suburban Background 	 -1.185319

60.13922 	 Shetland Islands 	 25/05/2005 	 NA 	 LH 	 Lullington Heath 	 Rural Background

0.181250 	 50.79370 	 Wealden District 	 29/09/1988 	 NA 	 LINC 	 Lincoln Roadside

Urban traffic 	 -0.537895 	 53.22889 	 Lincoln City Council 	 06/05/1997 	 22/12/1999 	 LIVR

Liverpool Centre 	 Urban Background 	 -2.980249 	 53.40845 	 Liverpool City Council 	 23/04/1993 	 23/09/2002

LN 	 Lough Navar 	 Rural Background 	 -7.900328 	 54.43951 	 Fermanagh District Council 	 02/10/1996

NA 	 LON6 	 London Eltham 	 Suburban Background 	 0.070842 	 51.45266 	 London Borough of Greenwich

08/10/1993 	 31/12/2000 	 LON7 	 London Eltham (HC) 	 Urban Background 	 0.070842 	 51.45266

London Borough of Greenwich 	 17/10/2003 	 NA 	 LV6 	 Liverpool Queen's Drive Roadside 	 Urban traffic 	 -2.962500

53.44694 	 Liverpool City Council 	 01/01/2008 	 NA 	 LVP 	 Liverpool Speke 	 Urban Background

-2.844333 	 53.34633 	 Liverpool City Council 	 24/11/1995 	 31/12/2000 	 LW1 	 London Lewisham

Urban Background 	 -0.020139 	 51.44541 	 London Borough of Lewisham 	 16/04/1997 	 30/09/2007 	 MACK

Charlton Mackrell 	 Rural Background 	 -2.683450 	 51.05625 	 South Somerset District 	 03/09/2008 	 NA

MAN 	 Manchester Town Hall 	 Urban Background 	 -2.244800 	 53.47850 	 Manchester City Council 	 08/08/1991

30/09/2007 	 MAN3 	 Manchester Piccadilly 	 Urban Background 	 -2.237881 	 53.48152 	 Manchester City Council

18/12/1995 	 30/09/2007 	 MAN4 	 Manchester South 	 Suburban Background 	 -2.243280 	 53.36903

Manchester City Council 	 06/12/1996 	 NA 	 MH 	 Mace Head 	 Rural Background 	 -9.903917

53.32644 	 Unknown 	 03/04/1987 	 NA 	 MID 	 Middlesbrough 	 Urban Background

-1.220874 	 54.56930 	 Middlesbrough 	 01/01/1993 	 31/12/2000 	 MKTH 	 Market Harborough

Rural Background 	 -0.772222 	 52.55444 	 Harborough District 	 10/12/2003 	 NA 	 MOLD

Mold 	 Suburban Background 	 -3.144889 	 53.16231 	 Flintshire County 	 02/12/2009 	 NA

MY1 	 London Marylebone Road 	 Urban traffic 	 -0.154611 	 51.52253 	 City of Westminster 	 01/04/1998

NA 	 NCA3 	 Newcastle Cradlewell Roadside 	 Urban traffic 	 -1.598611 	 54.98639 	 Newcastle City Council

10/03/2008 	 NA 	 NEWC 	 Newcastle Centre 	 Urban Background 	 -1.610528 	 54.97825

Newcastle City Council 	 08/03/1992 	 NA 	 NO10 	 Norwich Forum Roadside 	 Urban traffic 	 1.291714

52.62817 	 Norwich City Council 	 08/04/2005 	 30/09/2007 	 NO12 	 Norwich Lakenfields 	 Urban Background

1.301976 	 52.61419 	 Norwich City Council 	 01/09/2009 	 NA 	 NOR1 	 Norwich Roadside

Urban traffic 	 1.299064 	 52.62200 	 Norwich City Council 	 21/06/1997 	 14/02/2005 	 NOR2

Norwich Centre 	 Urban Background 	 1.295019 	 52.63203 	 Norwich City Council 	 24/07/1997 	 30/09/2007

NOTT 	 Nottingham Centre 	 Urban Background 	 -1.146447 	 52.95473 	 Nottingham City Council 	 02/09/1996

30/09/2007 	 NPT3 	 Newport 	 Urban Background 	 -2.977281 	 51.60120 	 Newport County Borough

01/01/2008 	 NA 	 NTON 	 Northampton 	 Urban Background 	 -0.885933 	 52.27349

Northampton District 	 12/03/2002 	 30/09/2007 	 OLDB 	 Sandwell Oldbury 	 Urban Background 	 -2.017629

52.50431 	 Sandwell District 	 27/06/1997 	 21/09/1998 	 OSY 	 St Osyth 	 Rural Background

1.049031 	 51.77798 	 Tendring District 	 11/05/2002 	 31/12/2009 	 OX 	 Oxford Centre Roadside

Urban traffic 	 -1.257581 	 51.75181 	 Oxford City Council 	 15/04/1996 	 30/09/2007 	 OX8

Oxford St Ebbes 	 Urban Background 	 -1.260278 	 51.74481 	 Oxford City Council 	 18/12/2008 	 NA

PEEB 	 Peebles 	 Urban Background 	 -3.196527 	 55.65747 	 Scottish Borders 	 06/11/2009

NA 	 PEMB 	 Narberth 	 Rural Background 	 -4.691462 	 51.78178 	 Pembrokshire County

20/01/1997 	 NA 	 PLYM 	 Plymouth Centre 	 Urban Background 	 -4.142361 	 50.37167

Plymouth City Council 	 29/09/1997 	 30/09/2007 	 PMTH 	 Portsmouth 	 Urban Background 	 -1.068583

50.82881 	 Portsmouth City Council 	 21/03/2002 	 30/09/2007 	 PRES 	 Preston 	 Urban Background

-2.680353 	 53.76559 	 Preston District 	 06/06/2000 	 30/09/2007 	 PT 	 Port Talbot

Urban Industrial 	 -3.761690 	 51.57980 	 Neath Port Talbot County Borough 	 09/01/1997 	 24/07/2007 	 PT4

Port Talbot Margam 	 Urban Industrial 	 -3.770822 	 51.58395 	 Neath Port Talbot County Borough 	 23/04/2008 	 NA

REA1 	 Reading New Town 	 Urban Background 	 -0.944067 	 51.45309 	 Reading 	 17/10/2003

30/09/2007 	 READ 	 Reading 	 Urban Background 	 -0.955180 	 51.45352 	 Reading

17/07/1997 	 06/02/2003 	 REDC 	 Redcar 	 Suburban Background 	 -1.073300 	 54.61073

Redcar and Cleveland 	 25/06/1997 	 30/09/2007 	 ROCH 	 Rochester Stoke 	 Rural Background 	 0.634889

51.45617 	 Medway 	 26/01/1996 	 NA 	 ROTH 	 Rotherham Centre 	 Urban Background

-1.354444 	 53.43186 	 Rotherham District 	 20/06/1997 	 30/09/2007 	 RUGE 	 Rugeley

Urban Background 	 -1.936111 	 52.75306 	 Cannock Chase District 	 21/03/1991 	 17/09/1992 	 SALT

Saltash Roadside 	 Urban traffic 	 -4.230300 	 50.41310 	 Caradon District 	 23/02/2010 	 31/08/2010

SCN2 	 Scunthorpe Town 	 Urban Industrial 	 -0.636811 	 53.58634 	 North Lincolnshire 	 10/01/2008

NA 	 SCUN 	 Scunthorpe 	 Urban Industrial 	 -0.633015 	 53.58499 	 North Lincolnshire

15/12/1997 	 18/03/2004 	 SDY 	 Sandy Roadside 	 Urban traffic 	 -0.300306 	 52.13242

Mid Bedfordshire District 	 28/07/2008 	 NA 	 SEND 	 Southend-on-Sea 	 Urban Background 	 0.678408

51.54421 	 Southend-on-Sea 	 24/07/2000 	 30/01/2009 	 SHE 	 Sheffield Tinsley 	 Urban Background

-1.396139 	 53.41058 	 Sheffield City Council 	 08/08/1991 	 30/09/2007 	 SHE2 	 Sheffield Centre

Urban Background 	 -1.473306 	 53.37772 	 Sheffield City Council 	 19/06/2007 	 NA 	 SIB

Sibton 	 Rural Background 	 1.463497 	 52.29440 	 Suffolk Coastal District 	 10/05/1977 	 31/03/1989

SK1 	 London Southwark 	 Urban Background 	 -0.096667 	 51.49055 	 London Borough of Southwark 	 28/02/1997

30/09/2007 	 SK2 	 Southwark Roadside 	 Urban traffic 	 -0.062300 	 51.48199 	 London Borough of Southwark

01/04/1997 	 09/02/2006 	 SK5 	 Southwark A2 Old Kent Road 	 Urban traffic 	 -0.059550 	 51.48050

London Borough of Southwark 	 24/10/2010 	 NA 	 SOM 	 Somerton 	 Rural Background 	 -2.735253

51.03572 	 South Somerset District 	 28/04/2003 	 05/03/2008 	 SOUT 	 Southampton Centre 	 Urban Background

-1.395778 	 50.90814 	 Southampton City Council 	 04/01/1994 	 NA 	 STE 	 Stevenage

Suburban Background 	 -0.200833 	 51.88694 	 Stevenage District 	 01/04/1977 	 29/04/1994 	 STEW

Stewartby 	 Urban Industrial 	 -0.511111 	 52.07194 	 Bedford District 	 26/11/2007 	 31/12/2008

STK4 	 Stockport Shaw Heath 	 Urban Background 	 -2.161111 	 53.40306 	 Stockport District 	 09/10/2002

30/09/2007 	 STOC 	 Stockport 	 Urban Background 	 -2.158200 	 53.40994 	 Stockport District

25/11/1996 	 03/10/2002 	 STOK 	 Stoke-on-Trent Centre 	 Urban Background 	 -2.175133 	 53.02821

Stoke-on-trent City Council 	 12/06/2007 	 NA 	 STOR 	 Storrington Roadside 	 Urban traffic 	 -0.450878

50.91742 	 Horsham District 	 01/08/2009 	 NA 	 SUN2 	 Sunderland Silksworth 	 Urban Background

-1.406878 	 54.88361 	 Sunderland District 	 09/12/2004 	 NA 	 SUND 	 Sunderland

Urban Background 	 -1.380081 	 54.90611 	 Sunderland District 	 06/10/1992 	 30/09/2007 	 SUT1

Sutton Roadside 	 Urban traffic 	 -0.182789 	 51.36636 	 London Borough of Sutton 	 01/04/1996 	 02/05/2002

SUT3 	 London Sutton 	 Suburban Background 	 -0.165489 	 51.36789 	 London Borough of Sutton 	 01/04/1996

02/05/2002 	 SV 	 Strath Vaich 	 Rural Background 	 -4.776583 	 57.73446 	 Highland

21/10/1990 	 14/03/1997 	 SWA1 	 Swansea Roadside 	 Urban traffic 	 -3.947374 	 51.63270

City and County of Swansea 	 20/09/2006 	 30/09/2007 	 SWAN 	 Swansea 	 Urban Background 	 -3.943329

51.62114 	 City and County of Swansea 	 01/12/1994 	 07/08/2006 	 TED 	 London Teddington 	 Urban Background

-0.339647 	 51.42099 	 London Borough of Richmond Upon Thames 	 08/08/1996 	 NA 	 TH2 	 Tower Hamlets Roadside

Urban traffic 	 -0.042164 	 51.52256 	 London Borough of Tower Hamlets 	 01/04/1996 	 NA 	 THUR

Thurrock 	 Urban Background 	 0.317969 	 51.47707 	 Thurrock 	 20/03/2009 	 NA

TRAN 	 Wirral Tranmere 	 Urban Background 	 -3.022722 	 53.37287 	 Wirral District 	 14/05/2000

30/09/2007 	 WA2 	 London Wandsworth 	 Urban Background 	 -0.191164 	 51.45696 	 London Borough of Wandsworth

01/04/1996 	 30/09/2007 	 WAL 	 Walsall Alumwell 	 Urban Background 	 -2.010483 	 52.58167

Walsall District 	 05/03/1987 	 30/09/2007 	 WAL2 	 Walsall Willenhall 	 Urban Background 	 -2.033144

52.60821 	 Walsall District 	 29/04/1997 	 03/02/2010 	 WAR 	 Warrington 	 Urban Background

-2.615358 	 53.38928 	 Warrington 	 21/10/2008 	 NA 	 WBRO 	 Sandwell West Bromwich

Urban Background 	 -1.995556 	 52.52062 	 Sandwell District 	 04/11/1998 	 30/09/2007 	 WC

Wharleycroft 	 Rural Background 	 -2.469167 	 54.61655 	 Eden District 	 08/05/1985 	 28/11/1995

WEYB 	 Weybourne 	 Rural Background 	 1.122017 	 52.95049 	 North Norfolk District 	 30/05/2001

NA 	 WFEN 	 Wicken Fen 	 Rural Background 	 0.290917 	 52.29850 	 East Cambridgeshire District

12/11/1997 	 NA 	 WIG3 	 Wigan Leigh 	 Urban Background 	 -2.506899 	 53.49422

Wigan District 	 15/05/2002 	 28/09/2004 	 WIG5 	 Wigan Centre 	 Urban Background 	 -2.638139

53.54914 	 Wigan District 	 08/10/2004 	 NA 	 WL 	 West London 	 Urban Background

-0.200361 	 51.49380 	 Royal Borough of Kensington and Chelsea 	 14/08/1989 	 30/09/2007 	 WOLV 	 Wolverhampton Centre

Urban Background 	 -2.129008 	 52.58818 	 Wolverhampton MBC 	 13/06/2007 	 30/09/2007 	 WRAY

Wray 	 Rural Background 	 -2.583333 	 54.10444 	 Lancaster City Council 	 01/04/1985 	 29/02/1988

WREX 	 Wrexham 	 Urban traffic 	 -3.002778 	 53.04222 	 Wrecsam - Wrexham 	 06/03/2002

30/09/2007 	 YARM 	 Stockton-on-Tees Yarm 	 Urban traffic 	 -1.354319 	 54.50918 	 Stockton-on-Tees

13/08/2002 	 30/09/2007 	 YK10 	 York Bootham 	 Urban Background 	 -1.086514 	 53.96751

City of York 	 10/02/2009 	 NA 	 YK11 	 York Fishergate 	 Urban traffic 	 -1.075861

53.95189 	 City of York 	 01/01/2008 	 NA 	 A3 	 London A3 Roadside 	 Urban traffic

See also

importKCL, importADMS, importSAQN

Examples

## import all pollutants from Marylebone Rd from 1990:2009
## Not run: mary &lt;- importAURN(site = &quot;my1&quot;, year = 2000:2009)

## import nox, no2, o3 from Marylebone Road and Nottingham Centre for 2000
## Not run: ------------------------------------
# thedata <- importAURN(site = c("my1", "nott"), year = 2000,
# pollutant = c("nox", "no2", "o3"))
## ---------------------------------------------

## import over 20 years of Mace Head O3 data!
## Not run: o3 &lt;- importAURN(site = &quot;mh&quot;, year = 1987:2009)

## import hydrocarbon (and other) data from Marylebone Road
## Not run: mary &lt;- importAURN(site = &quot;my1&quot;, year =1998, hc = TRUE)

## reshape the data so that each column represents a pollutant/site
## Not run: ------------------------------------
# require(reshape2)
# thedata <- importAURN(site = c("nott", "kc1"), year = 2008,
# pollutant = "o3")
# thedata <- melt(thedata, measure.vars = "o3")
# thedata <- dcast(thedata, ... ~ variable + site + code)
# ## thedata now has columns  o3_Nottingham Centre_NOTT o3_London N. Kensington_KC1
# 
## ---------------------------------------------