Import data from the UK Air Pollution Networks

Functions for importing air pollution data from a range of UK networks including the Automatic Urban and Rural Network (AURN), the individual England (AQE), Scotland (SAQN), Wales (WAQN) and Northern Ireland (NI) Networks, and many "locally managed" monitoring networks across England. Files are imported from a remote server operated by Ricardo that provides air quality data files as R data objects.

Usage

importUKAQ(
  site = "my1",
  year = 2022,
  source = "aurn",
  data_type = "hourly",
  pollutant = "all",
  hc = FALSE,
  meta = FALSE,
  meteo = TRUE,
  ratified = FALSE,
  to_narrow = FALSE,
  verbose = FALSE,
  progress = TRUE
)

Arguments

site

Site code of the site to import, e.g., "my1" is Marylebone Road. Site codes can be discovered through the use of importMeta(). Several sites can be imported at once. For example, site = c("my1", "nott") imports both Marylebone Road and Nottingham. Sites from different networks can be imported through also providing multiple sources.

year

Year(s) to import. To import a series of years use, e.g., 2000:2020. To import several specific years use year = c(2000, 2010, 2020).

source

The network to which the site(s) belong, defaulting to "aurn". Providing a single network will attempt to import all of the given sites from the provided network. Alternatively, a vector of sources can be provided of the same length as site to indicate which network each site individually belongs. Available networks include:

"aurn", The UK Automatic Urban and Rural Network.

"aqe", The Air Quality England Network.

"saqn", The Scottish Air Quality Network.

"waqn", The Welsh Air Quality Network.

"ni", The Northern Ireland Air Quality Network.

"local", Locally managed air quality networks in England.

data_type

The type of data to be returned, defaulting to "hourly" data. Alternative data types include:

"daily": Daily average data.

"monthly": Monthly average data with data capture information for the whole network.

"annual": Annual average data with data capture information for the whole network.

"15_min": 15-minute average SO2 concentrations.

"8_hour": 8-hour rolling mean concentrations for O3 and CO.

"24_hour": 24-hour rolling mean concentrations for particulates.

"daily_max_8": Maximum daily rolling 8-hour maximum for O3 and CO.

"daqi": Daily Air Quality Index (DAQI). See here for more details of how the index is defined. Note that this data_type is not available for locally managed monitoring networks.

pollutant

Pollutants to import. If omitted will import all pollutants from a site. To import only NOx and NO2 for example use pollutant = c("nox", "no2").

hc

Include hydrocarbon measurements in the imported data? Defaults to FALSE as most users will not be interested in using hydrocarbon data.

meta

Append the site type, latitude and longitude of each selected site? Defaults to FALSE.

meteo

Append modelled meteorological data, if available? Defaults to TRUE, which will return wind speed (ws), wind direction (wd) and ambient temperature (air_temp). The variables are calculated from using the WRF model run by Ricardo Energy & Environment and are available for most but not all networks. Setting meteo = FALSE is useful if you have other meteorological data to use in preference, for example from worldmet.

ratified

Append qc column(s) to hourly data indicating whether each species was ratified (i.e., quality-checked)? Defaults to FALSE.

to_narrow

Return the data in a "narrow"/"long"/"tidy" format? By default the returned data is "wide" and has a column for each pollutant/variable. When to_narrow = TRUE the data are returned with a column identifying the pollutant name and a column containing the corresponding concentration/statistic. Defaults to FALSE.

verbose

Print messages to the console if hourly data cannot be imported? Default is FALSE. TRUE is useful for debugging as the specific year(s), site(s) and source(s) which cannot be imported will be returned.

progress

Show a progress bar when many sites/years are being imported? Defaults to TRUE.

Value

a tibble

Importing UK Air Pollution Data

This family of functions has been written to make it easy to import data from across several UK air quality networks. Ricardo have provided .RData files (R workspaces) of all individual sites and years, as well as up to date meta data. These files are updated on a daily basis. This approach requires a link to the Internet to work.

For an up to date list of available sites that can be imported, see importMeta().

The site codes and pollutant names can be upper or lower case.

There are several advantages over the web portal approach where .csv files are downloaded. First, it is quick to select a range of sites, pollutants and periods (see examples below). Second, storing the data as .RData objects is very efficient as they are about four times smaller than .csv files --- which means the data downloads quickly and saves bandwidth. Third, the function completely avoids any need for data manipulation or setting time formats, time zones etc. The function also has the advantage that the proper site name is imported and used in {openair} functions.

The data are imported by stacking sites on top of one another and will have field names site, code (the site code) and pollutant.

By default, the function returns hourly average data. However, annual, monthly, daily and 15 minute data (for SO2) can be returned using the option data_type. Annual and monthly data provide whole network information including data capture statistics.

All units are expressed in mass terms for gaseous species (ug/m3 for NO, NO2, NOx (as NO2), SO2 and hydrocarbons; and mg/m3 for CO). PM10 concentrations are provided in gravimetric units of ug/m3 or scaled to be comparable with these units. Over the years a variety of instruments have been used to measure particulate matter and the technical issues of measuring PM10 are complex. In recent years the measurements rely on FDMS (Filter Dynamics Measurement System), which is able to measure the volatile component of PM. In cases where the FDMS system is in use there will be a separate volatile component recorded as 'v10' and non-volatile component 'nv10', which is already included in the absolute PM10 measurement. Prior to the use of FDMS the measurements used TEOM (Tapered Element Oscillating. Microbalance) and these concentrations have been multiplied by 1.3 to provide an estimate of the total mass including the volatile fraction.

Some sites report hourly and daily PM10 and / or PM2.5. When data_type = "daily" and there are both hourly and 'proper' daily measurements available, these will be returned as e.g. "pm2.5" and "gr_pm2.5"; the former corresponding to data based on original hourly measurements and the latter corresponding to daily gravimetric measurements.

The function returns modelled hourly values of wind speed (ws), wind direction (wd) and ambient temperature (air_temp) if available (generally from around 2010). These values are modelled using the WRF model operated by Ricardo.

The BAM (Beta-Attenuation Monitor) instruments that have been incorporated into the network throughout its history have been scaled by 1.3 if they have a heated inlet (to account for loss of volatile particles) and 0.83 if they do not have a heated inlet. The few TEOM instruments in the network after 2008 have been scaled using VCM (Volatile Correction Model) values to account for the loss of volatile particles. The object of all these scaling processes is to provide a reasonable degree of comparison between data sets and with the reference method and to produce a consistent data record over the operational period of the network, however there may be some discontinuity in the time series associated with instrument changes.

No corrections have been made to the PM2.5 data. The volatile component of FDMS PM2.5 (where available) is shown in the 'v2.5' column.

See also

Author

David Carslaw, Trevor Davies, and Jack Davison

Examples

# import a single site from the AURN
importUKAQ("my1", year = 2022)
#> # A tibble: 8,760 × 15
#>    source site     code  date                   co   nox   no2    no    o3   so2
#>    <chr>  <chr>    <chr> <dttm>              <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 aurn   London … MY1   2022-01-01 00:00:00 0.244  65.4  33.3  21.0  17.2 0.266
#>  2 aurn   London … MY1   2022-01-01 01:00:00 0.338 117.   43.6  47.6  13.0 1.33 
#>  3 aurn   London … MY1   2022-01-01 02:00:00 0.291  76.7  33.7  28.1  18.8 0.532
#>  4 aurn   London … MY1   2022-01-01 03:00:00 0.291  90.8  37.5  34.7  21.0 0.798
#>  5 aurn   London … MY1   2022-01-01 04:00:00 0.244  93.7  33.1  39.5  20.2 0.798
#>  6 aurn   London … MY1   2022-01-01 05:00:00 0.210  85.3  34.2  33.4  21.8 0.798
#>  7 aurn   London … MY1   2022-01-01 06:00:00 0.186  63.7  27.3  23.7  21.0 0.266
#>  8 aurn   London … MY1   2022-01-01 07:00:00 0.210  67.9  31.2  23.9  21.0 0.532
#>  9 aurn   London … MY1   2022-01-01 08:00:00 0.198  63.1  30.2  21.6  20.8 0.532
#> 10 aurn   London … MY1   2022-01-01 09:00:00 0.210  73.8  33.3  26.3  23.0 0.532
#> # ℹ 8,750 more rows
#> # ℹ 5 more variables: pm10 <dbl>, pm2.5 <dbl>, ws <dbl>, wd <dbl>,
#> #   air_temp <dbl>

# import sites from another network
importUKAQ(c("bn1", "bn2"), year = 2022, source = "aqe")
#> Importing AQ Data ■■■■■■■■■■■■■■■■                  50% |  ETA:  2s
#> # A tibble: 17,520 × 8
#>    source site            code  date                  nox   no2    no  pm10
#>    <chr>  <chr>           <chr> <dttm>              <dbl> <dbl> <dbl> <dbl>
#>  1 aqe    Barnet Tally Ho BN1   2022-01-01 00:00:00 33.9  21.4  8.15   32.2
#>  2 aqe    Barnet Tally Ho BN1   2022-01-01 01:00:00 15.3  10.1  3.43   22.7
#>  3 aqe    Barnet Tally Ho BN1   2022-01-01 02:00:00 14.2   9.10 3.30   17  
#>  4 aqe    Barnet Tally Ho BN1   2022-01-01 03:00:00  9.32  6.69 1.71   16.7
#>  5 aqe    Barnet Tally Ho BN1   2022-01-01 04:00:00  7.39  6.11 0.839  17.1
#>  6 aqe    Barnet Tally Ho BN1   2022-01-01 05:00:00 17.0  11.0  3.87   16.9
#>  7 aqe    Barnet Tally Ho BN1   2022-01-01 06:00:00 20.8  12.1  5.66   16.3
#>  8 aqe    Barnet Tally Ho BN1   2022-01-01 07:00:00 21.9  14.8  4.68   11.3
#>  9 aqe    Barnet Tally Ho BN1   2022-01-01 08:00:00 22.7  14.7  5.22   11.9
#> 10 aqe    Barnet Tally Ho BN1   2022-01-01 09:00:00 30.0  17.7  7.98   18.2
#> # ℹ 17,510 more rows

# import sites across multiple networks
importUKAQ(c("my1", "bn1", "bn2"),
  year = 2022,
  source = c("aurn", "aqe", "aqe")
)
#> Importing AQ Data ■■■■■■■■■■■                       33% |  ETA:  3s
#> Importing AQ Data ■■■■■■■■■■■■■■■■■■■■■             67% |  ETA:  2s
#> # A tibble: 26,280 × 15
#>    source site     code  date                  nox   no2    no  pm10    co    o3
#>    <chr>  <chr>    <chr> <dttm>              <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 aqe    Barnet … BN1   2022-01-01 00:00:00 33.9  21.4  8.15   32.2    NA    NA
#>  2 aqe    Barnet … BN1   2022-01-01 01:00:00 15.3  10.1  3.43   22.7    NA    NA
#>  3 aqe    Barnet … BN1   2022-01-01 02:00:00 14.2   9.10 3.30   17      NA    NA
#>  4 aqe    Barnet … BN1   2022-01-01 03:00:00  9.32  6.69 1.71   16.7    NA    NA
#>  5 aqe    Barnet … BN1   2022-01-01 04:00:00  7.39  6.11 0.839  17.1    NA    NA
#>  6 aqe    Barnet … BN1   2022-01-01 05:00:00 17.0  11.0  3.87   16.9    NA    NA
#>  7 aqe    Barnet … BN1   2022-01-01 06:00:00 20.8  12.1  5.66   16.3    NA    NA
#>  8 aqe    Barnet … BN1   2022-01-01 07:00:00 21.9  14.8  4.68   11.3    NA    NA
#>  9 aqe    Barnet … BN1   2022-01-01 08:00:00 22.7  14.7  5.22   11.9    NA    NA
#> 10 aqe    Barnet … BN1   2022-01-01 09:00:00 30.0  17.7  7.98   18.2    NA    NA
#> # ℹ 26,270 more rows
#> # ℹ 5 more variables: so2 <dbl>, pm2.5 <dbl>, ws <dbl>, wd <dbl>,
#> #   air_temp <dbl>

# get "long" format hourly data with a ratification flag
importUKAQ(
  "card",
  source = "waqn",
  year = 2022,
  to_narrow = TRUE,
  ratified = TRUE
)
#> # A tibble: 70,080 × 10
#>    source date                site   code     ws    wd air_temp pollutant  value
#>    <chr>  <dttm>              <chr>  <chr> <dbl> <dbl>    <dbl> <chr>      <dbl>
#>  1 waqn   2022-01-01 00:00:00 Cardi… CARD    5.6  183.     10.6 co        0.0928
#>  2 waqn   2022-01-01 01:00:00 Cardi… CARD    8.3  198.     11.3 co        0.0843
#>  3 waqn   2022-01-01 02:00:00 Cardi… CARD    9.2  197      11.4 co        0.0854
#>  4 waqn   2022-01-01 03:00:00 Cardi… CARD    9.8  189.     11.3 co        0.0885
#>  5 waqn   2022-01-01 04:00:00 Cardi… CARD    9.2  188.     11.2 co        0.0949
#>  6 waqn   2022-01-01 05:00:00 Cardi… CARD    8.2  184.     11.1 co        0.0790
#>  7 waqn   2022-01-01 06:00:00 Cardi… CARD    7.6  178.     11.1 co        0.0632
#>  8 waqn   2022-01-01 07:00:00 Cardi… CARD    5.9  176.     11.2 co        0.0664
#>  9 waqn   2022-01-01 08:00:00 Cardi… CARD    4.4  174.     11   co        0.0601
#> 10 waqn   2022-01-01 09:00:00 Cardi… CARD    3.9  185.     10.7 co        0.0696
#> # ℹ 70,070 more rows
#> # ℹ 1 more variable: qc <lgl>

# import other data types, filtering by pollutant
importUKAQ(
  data_type = "annual",
  pollutant = c("no2", "pm2.5", "pm10"),
  source = c("aurn", "aqe")
)
#> # A tibble: 335 × 11
#>    source uka_code code  site        date                  no2 no2_capture  pm10
#>    <chr>  <fct>    <chr> <chr>       <dttm>              <dbl>       <dbl> <dbl>
#>  1 aurn   UKA00513 ABD7  Aberdeen U… 2022-01-01 00:00:00 26.6        0.994 NA   
#>  2 aurn   UKA00615 ABD8  Aberdeen W… 2022-01-01 00:00:00 24.3        0.997 NA   
#>  3 aurn   UKA00933 ABD9  Aberdeen E… 2022-01-01 00:00:00 16.5        0.995 11.3 
#>  4 aurn   UKA00451 ACTH  Auchencort… 2022-01-01 00:00:00 NA         NA      6.23
#>  5 aurn   UKA00137 AH    Aston Hill  2022-01-01 00:00:00  3.20       0.971 NA   
#>  6 aurn   UKA00541 ARM6  Armagh Roa… 2022-01-01 00:00:00 22.2        0.987 16.4 
#>  7 aurn   UKA00633 BAAR  Ballymena … 2022-01-01 00:00:00 15.8        0.914 NA   
#>  8 aurn   UKA00503 BALM  Ballymena … 2022-01-01 00:00:00 10.4        0.912 NA   
#>  9 aurn   UKA00353 BAR3  Barnsley G… 2022-01-01 00:00:00 12.8        0.952 NA   
#> 10 aurn   UKA00625 BBRD  Birkenhead… 2022-01-01 00:00:00 16.7        0.906 NA   
#> # ℹ 325 more rows
#> # ℹ 3 more variables: pm10_capture <dbl>, pm2.5 <dbl>, pm2.5_capture <dbl>