polar.annulus(openair)	R Documentation

Bivariate polar annulus plot

Description

Typically plots the concentration of a pollutant by wind direction and as a function of time as an annulus. The function is good for visualising how concentrations of pollutants vary by wind direction and a time period e.g. by month, day of week.

Usage


polar.annulus(polar,
pollutant = "",
resolution = "fine",
local.time = TRUE,
type = "trend",
limits = c(0, 100),
cols = "default",
width = "normal",
auto.smooth = FALSE,
k = 15, key = TRUE,
main = "",
auto.text = TRUE, ...)

Arguments

polar 	 A data frame minimally containing ws, wd and a pollutant. Can also contain date if plots by time period are required.

pollutant 	 Mandatory. A pollutant or other variable name corresponding to a variable in a data frame should be supplied e.g. pollutant = "nox"

resolution 	 Two plot resolutions can be set: normal and fine (the default).

local.time 	 Should the results be calculated in local time? The default is TRUE. Emissions activity tends to occur at local time e.g. rush hour is at 8 am every day. When the clocks go forward in spring, the emissions are effectively released into the atmosphere at GMT – 1 hour during the summer. When plotting diurnal profiles, this has the effect of "smearing-out" the concentrations. A better approach is to express time as local time, which here is defined as BST (British Summer Time). This correction tends to produce better-defined diurnal profiles of concentration (or other variables) and allows a better comparison to be made with emissions/activity data. If set to FALSE then GMT is used.

type 	 This determines the temporal period to consider. Options are "trend" (the default), "season" to plot variation throughout the year, "weekday" to plot day of the week variation and "hour" to plot the diurnal variation.

limits 	 The functions does its best to choose sensible limits automatically. However, there are circumstances when the user will wish to set different ones. An example would be a series of plots showing each year of data separately. The limits are set in the form c(lower, upper), so limits = c(0, 100) would force the plot limits to span 0–100.

cols 	 Colours to be used for plotting. Options include "default", "increment", "heat", "jet" and user defined. For user defined the user can supply a list of colour names recognised by R (type colours() to see the full list). An example would be cols = c("yellow", "green", "blue")

width 	 The width of the annulus; can be "normal" (the default), "thin" or "fat".

auto.smooth 	 Should smoothing be automatically controlled? The default is FALSE. When set to TRUE, this means that the GAM smoothing tries to find the best fit to the surface i.e. a compromise between a surface that is too noisy and too smooth. See mgcv for more details. On most occasions it is probably better to set it to FALSE and set the smooth parameter k separately; see next option.

k 	 This is the smoothing parameter that is set if auto.smooth is set to FALSE. Typically, value of around 15 (the default) seems to be suitable and will resolve more features in the plot.

exclude.missing 	 Setting this option to TRUE (the default) removes points from the plot that are too far from the original data. The smoothing routines will produce predictions at points where no data exist i.e. they predict. By removing the points too far from the original data produces a plot where it is clear where the original data lie. If set to FALSE missing data will be interpolated.

key 	 Should a key be drawn?

main 	 Title of plot.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the ‘2’ in NO2.

... 	 Other graphical parameters.

Details

The polar.annulus function shares many of the properties of the polar.plot. However, polar.annulus is focussed on displaying information on how concentrations of a pollutant (values of another variable) vary with wind direction and time. Plotting as an annulus helps to reduce compression of information towards the centre of the plot. The circular plot is easy to interpret because wind direction is most easily understood in polar rather than cartesian coordinates.

The inner part of the annulus represents the earliest time and the outer part of the annulus the latest time. The time dimension can be shown in many ways including "trend" (default), "hour" (hour or day), "season" (month of the year) and "weekday" (day of the week). Taking hour as an example, the plot will show how concentrations vary by hour of the day and wind direction. Such plots can be very useful for understanding how different source influences affect a location.

Warning

....

Author(s)

David Carslaw

See Also

polar.plot, polar.freq

Examples



# load example data from package
data(mydata)

# trend plot for PM10 at Marylebone Rd
polar.annulus(mydata, poll="pm10", main = "trend in pm10 at Marylebone Road")

# seasonal plot for PM10 at Marylebone Rd
polar.annulus(mydata, poll="pm10", type = "season")

# trend in coarse particles (PMc = PM10 - PM2.5), calculate PMc first

mydata$pmc <- mydata$pm10 - mydata$pm25
polar.annulus(mydata, poll="pmc", type = "trend in pmc at Marylebone Road")


[Package openair version 1.0 Index]