polar.condition(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


polar.condition(polar, pollutant = "", resolution = "normal", limits = c(0, 100), exclude.missing = TRUE, cols = "jet", min.bin = 1, upper = 10, auto.smooth = FALSE, k = 100, xlim = c(-upper - 3, upper + 3), ylim = c(-upper - 3, upper + 3), ...)

Arguments

polar 	 ~~Describe polar here~~

pollutant 	 ~~Describe pollutant here~~

resolution 	 ~~Describe resolution here~~

limits 	 ~~Describe limits here~~

exclude.missing 	 ~~Describe exclude.missing here~~

cols 	 ~~Describe cols here~~

min.bin 	 ~~Describe min.bin here~~

upper 	 ~~Describe upper here~~

auto.smooth 	 ~~Describe auto.smooth here~~

k 	 ~~Describe k here~~

xlim 	 ~~Describe xlim here~~

ylim 	 ~~Describe ylim here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(polar,
              pollutant = "",
              resolution = "normal",
              limits = c(0,100),
              exclude.missing = TRUE,
              cols = "jet",
              min.bin = 1,
              upper = 10,
              auto.smooth = FALSE,
              k = 100,
              xlim = c(-upper - 3, upper + 3), 
              ylim =c(-upper - 3, upper + 3),...) {

  ## Input variables are described as follows:

  ## polar - data.frame consisting of ws, wd and conc.  This is the only
  ## compulsary requirement, all others below are optional and have defaults

  ## resolution - can be "normal" (default) or "fine".  Normal should be
  ## adequate for most purposes; "fine" doubles the resolution for smoother
  ## rendering.

  ## limits - the user can specify their own lower and upper limits if the
  ## default scale is not appropriate.  This is input in the form
  ## c(lower, upper).

  ## exlude.missing - if TRUE (default) plots are shown with missing data
  ## removed.



  #needs access to these packages
  library(mgcv)    #for smoothing surfaces and removing noise
  library(vcd)     #colours ("heat")
  library(matlab)  #colours ("jet")
  library(lattice) #basic plotting
  library(IDPmisc) #more colours ("default")

  #extract variables of interest
  vars <- c("ws", "wd", "date", pollutant)
  polar <- polar[, vars]
  polar <- polar[order(polar$date), ]

  #upper is useful for extending plots with ws to > 10 m/s

  #upper ws
  if (upper < 10) upper <- 10
  
  ws.scale <- 1
  d <- 0
  

  polar$wd[polar$wd == 0] <- 360

  #for resolution of grid plotting (default = 0.2; fine =0.1)
  if (resolution =="normal") int <- 0.2
  if (resolution =="fine") int <- 0.1
  if (resolution =="ultra.fine") int <- 0.05  #very large files!

  #binning wd data properly
  max.ws <- ceiling(max(polar$ws, na.rm = TRUE))
  ws <- seq(0, max.ws, length = 30)
  wd <- seq(from = 10, to = 360, by = 10) #wind directions from 10 to 360
  ws.wd <- expand.grid(ws = ws, wd = wd)

  #round wd to make processing obvious
  polar$wd[polar$wd == 0] <- NA
  polar$wd <- 10 * round(polar$wd/10)
  polar$wd[polar$wd == 0] <- 360   # angles <5 should be in 360 bin
  polar$wd[polar$wd == 0] <- 360
  
  u <- ws.wd$ws * sin(pi * ws.wd$wd / 180) #convert to polar coords
  v <- ws.wd$ws * cos(pi * ws.wd$wd / 180)

  #data to predict over
  input.data <- expand.grid(u = seq(-upper - d, upper + d, by = int),
                v = seq(-upper - d, upper + d, by = int))

prepare.grid <- function(polar) {
  #identify which ws and wd bins the data belong
  wd <- cut(polar$wd, breaks = seq(0, 360, 10), include.lowest = TRUE)
  ws <- cut(polar$ws, breaks = seq(0, max.ws, length = 31))

   #this automatically deals with missing data
  binned <- tapply(polar[, pollutant], list(wd, ws), mean, na.rm = TRUE)
  binned <- as.vector(t(binned))
  
  #frequency - remove points with freq < min.bin
  bin.len <- tapply(polar[, pollutant], list(wd, ws), length)
  binned.len <- as.vector(t(bin.len))
  ids <- which(binned.len < min.bin)
  binned[ids] <- NA
######################Smoothing#################################################

    #run GAM to make a smooth surface
    #use square root to avoid any problems with negative predictions (rare)
    if (auto.smooth == TRUE) {
        Mgam <- gam(binned ~ s(u, v))
     } else {
        Mgam <- gam(binned ~ s(u, v, k = k))
    }
    pred <- predict.gam(Mgam, input.data)
    pred <- pred

  #############################################################################

  results <- data.frame(u = input.data$u, v = input.data$v, z = pred)

  if (exclude.missing == TRUE) {
      # exclude predictions too far from data (from mgcv)
      # assumes wind speeds up to 10
      x <- seq(-upper, upper, by = int)
      y <- x
      res <- (2*upper)/int + 1
      wsp <- rep(x, res)
      wdp <- rep(y, rep(res, res))
      #remove null data
      polar <- na.omit(polar)

      ind <- exclude.too.far(wsp, wdp, polar$ws * sin(pi * polar$wd/180),
             polar$ws * cos(pi * polar$wd/180), dist = 0.05)

      results$z[ind] <- NA
      results
  }
  }

#############################################################################
  #cut data by year
  polar$cond <- cut(polar$date, "year", labels = FALSE)
  polar <- na.omit(polar)
  #prepare input grid
  results.grid <- data.frame(u = NULL, v = NULL, z = NULL, cond = NULL)

  for (i in unique(polar$cond)) {
      pred <- prepare.grid(subset(polar, cond == i))
      pred <- cbind(pred, cond = i)
      results.grid <- rbind(results.grid, pred)
  }

  #remove wind speeds > upper to make a circle
  results.grid$z[(results.grid$u^2 + results.grid$v^2)^.5 > upper*ws.scale] <- NA
  
  results.grid$cond <- as.factor(results.grid$cond)
  levels(results.grid$cond) = unique(format(polar$date, "%Y"))

  #auto-scaling
  nlev = 200  #preferred number of intervals
  #handle missing breaks arguments
    if(missing(limits)) {
          breaks=pretty(results.grid$z, n = nlev)
     } else {
          breaks=pretty(limits, n = nlev)
    }

  nlev2 = length(breaks)

  #various colour schemes ######################################################

  if(cols == "default") col <- IDPcolorRamp(nlev2 - 1)
  if(cols == "jet") col <- jet.colors(nlev2 - 1)
  if(cols == "heat") col <- rev(heat_hcl(nlev2 - 1, c = c(100, 30),
            l = c(30, 100), power = c(1/5,3)))
  ##############################################################################

  col.scale = breaks

contourplot(z ~ u * v | cond, results.grid, axes = FALSE,
      as.table = TRUE,
      col.regions = col, region = TRUE,
      aspect = 1, at = col.scale, xlab = "", ylab = "",
      scales = list(draw = FALSE),
      xlim = c(-upper - 3, upper + 3), ylim =c(-upper - 3, upper + 3),
      ...,

      #colorkey = list(labels = list(labels=as.character(col.scale),
      #  at=col.scale, cex = 0.7)),
      panel = function(x, y, z,subscripts,...) {
              panel.contourplot(x, y, z, subscripts, at = col.scale,
              pretty = TRUE,
              col.regions = col, labels = FALSE)

              #add axis line to central polar plot
              llines(c(-upper*ws.scale, upper*ws.scale), c(0, 0),
                      col = "grey20")
              llines(c(0, 0), c(-upper*ws.scale, upper*ws.scale),
                      col = "grey20")

               #annotate
               lsegments(-upper:upper,
                  rep(-.2, 2 * upper + 1),
                  -upper:upper,
                  rep(.2, 2 * upper + 1), col = "grey20")
               lsegments(rep(-.2, 2 * upper + 1),
                  -upper:upper,
                  rep(.2, 2 * upper + 1),
                  -upper:upper, col = "grey20")
                #larger ticks every 5
                lsegments(seq(-upper, upper, by = 5),
                  rep(-.4, 2 * upper + 1),
                  seq(-upper , upper, by = 5),
                  rep(.4, 2*upper + 1), col = "grey20")
               lsegments(rep(-.4, 2 * upper + 1),
                  seq(-upper, upper, by = 5),
                  rep(.4, 2 * upper + 1),
                  seq(-upper, upper, by = 5), col = "grey20")

        # text for directions

            ltext(-upper - 1.5, 0, "W")
            ltext(0, -upper - 1.5, "S")
            ltext(0, upper + 1.5, "N")
            ltext(upper + 1.5, 0, "E")
      })

  }

[Package openair version 1.0 Index]