polar.plot(openair)	R Documentation

Bivariate polar plot with smoothing

Description

Function for plotting pollutant concentration in polar coordinates showing concentration by wind speed and direction. Mean concentrations are calculated for wind speed-direction ‘bins’ (0-1, 1-2 m/s,... and 0-10, 10-20 degrees etc.). To aid interpretation, gam smoothing is carried out using mgcv.

Usage


polar.plot(polar,
pollutant = "nox",
type = "default",
resolution = "normal",
limits = c(0, 100),
exclude.missing = TRUE,
cols = "default",
min.bin = 1,
upper = 10,
force.positive = TRUE,
k = 100, main = "",
auto.text = TRUE, ...)  

Arguments

polar 	 A data frame minimally containing ws, wd and a pollutant. Can also contain date if plots by time period are required.

pollutant 	 Mandatory. A pollutant name corresponding to a variable in a data frame should be supplied e.g. pollutant = "nox"

type 	 The type of analysis to be done. The default is will produce a single plot using the entire data. Other types include "hour" (for hour of the day), "weekday" (for day of the week) and "month" (for month of the year), "year" for a polar plot for each year. There is also an option "site", which requires users to format their data in the form: date, ws, wd, site.It is also possible to choose type as another variable in the data frame. For example, type = "o3" will plot four wind roses for different levels of ozone, split into four quantiles (approximately equal numbers of counts in each of the four splits). This offers great flexibility for understanding the variation of different variables dependent on another.

resolution 	 Two plot resolutions can be set: normal (the default) and fine, for a smoother plot. It should be noted that plots with a fine resolution can take longer to render and the default option should be sufficient or most circumstances.

limits 	 The functions does its best to choose sensible limits automatically. However, there are circumstances when the user will wish to set different ones. An example would be a series of plots showing each year of data separately. The limits are set in the form c(lower, upper), so limits = c(0, 100) would force the plot limits to span 0-100.

exclude.missing 	 Setting this option to TRUE (the default) removes points from the plot that are too far from the original data. The smoothing routines will produce predictions at points where no data exist i.e. they predict. By removing the points too far from the original data produces a plot where it is clear where the original data lie. If set to FALSE missing data will be interpolated.

cols 	 Colours to be used for plotting. Options include "default", "increment", "heat", "jet" and user defined. For user defined the user can supply a list of colour names recognised by R (type colours() to see the full list). An example would be cols = c("yellow", "green", "blue")

min.bin 	 The minimum number of points allowed in a wind speed/wind direction bin. The default is 1. A value of two requires at least 2 valid records in each bin (units of 1 m/s and 10 degrees) an so on; bins with less than 2 valid records are set to NA. Care should be taken when using a value > 1 because of the risk of removing real data points. It is recommended to consider your data with care. Also, the polar.freq function can be of use in such circumstances.

upper 	 This sets the upper limit wind speed to be used. Often there are only a relatively few data points at very high wind speeds and plotting all of them can reduce the useful information in the plot.

force.positive 	 The default is TRUE. Sometimes if smoothing data with streep gradients it is possible for predicted values to be negative. force.positive = TRUE ensures that predictions remain postive. This is useful for several reasons. First, with lots of missing data more interpolation is needed and this can result in artefacts because the predictions are too far from the original data. Second, if it is known beforehand that the data are all postive, then this option carries that assumption through to the prediction. The only likley time where setting force.positive = FALSE would be if background concentrations were first subtracted resulting in data that is legitimately negative. For the vast majority of situations it is expected that the user will not need to alter the default option.

k 	 This is the smoothing parameter that is set if auto.smooth is set to FALSE. Typically, value of around 100 (the default) seems to be suitable and will resolve more features in the plot.

main 	 The plot title; default is no title.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the `2' in NO2.

... 	 Other graphical parameters.

Details

The bivariate polar plot is a useful diagnostic tool for quickly gaining an idea of potential sources. Wind speed is one of the most useful variables to use to separate source types (see references). For example, ground-level concentrations resulting from buoyant plumes from chimney stacks tend to peak under higher wind speed conditions. Conversely, ground-level, non-buoyant plumes such as from road traffic, tend to have highest concentrations under low wind speed conditions. Other sources such as from aircraft engines also show differing characteristics by wind speed.

The plots can vary considerably depending on how much smoothing is done. The approach adopted here is based on the very flexible and capable mgcv package that uses Generalized Additive Models. While methods do exist to find an optimum level of smoothness, they are not necessarily useful. The principal aim of polar.plot is as a graphical analysis rather than for quantitative purposes. In this respect the smoothing aims to strike a balance between interesting (real) features and overly noisy data. The defaults used in polar.plot are based on the analysis of data from many different sources. More advanced users may wish to modify the code and adopt other smoothing approaches.

Variants on polar.plot include polar.annulus and polar.freq.

Warning

Some plots can take a long time to produce e.g. if there are many years of data and type = "year" is chosen. The function may fail if there are insufficient data to smooth the surface.

Author(s)

David Carslaw

References

Carslaw, D.C., Beevers, S.D, Ropkins, K and M.C. Bell (2006). Detecting and quantifying aircraft and other on-airport contributions to ambient nitrogen oxides in the vicinity of a large international airport. Atmospheric Environment. 40/28 pp 5424-5434.

Henry, R.C., Chang, Y.S., Spiegelman, C.H., 2002. Locating nearby sources of air pollution by nonparametric regression of atmospheric concentrations on wind direction. Atmospheric Environment 36 (13), 2237-2244.

Westmoreland, E.J., N. Carslaw, D.C. Carslaw, A. Gillah and E. Bates (2007). Analysis of air quality within a street canyon using statistical and dispersion modelling techniques. Atmospheric Environment. Vol. 41(39), pp. 9195-9205.

Yu, K.N., Cheung, Y.P., Cheung, T., Henry, R.C., 2004. Identifying the impact of large urban airports on local air quality by nonparametric regression. Atmospheric Environment 38 (27), 4501-4507.

See Also

polar.annulus, polar.freq

Examples



# load example data from package
data(mydata)

# basic plot
polar.plot(mydata, pollutant = "nox")

# polar plots by year on same scale
## Not run: polar.plot(mydata, pollutant = "so2", type = "year", main = "polar plot of so2")

# set minimum number of bins to be used to see if pattern remains similar
polar.plot(mydata, pollutant = "nox", min.bin = 3)

# plot by day of the week

## Not run: polar.plot(mydata, pollutant = "pm10", type = "weekday")


[Package openair version 0.1 Index]