samp.boot.block(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


samp.boot.block(n, B, blockLength = 20)

Arguments

n 	 ~~Describe n here~~

B 	 ~~Describe B here~~

blockLength 	 ~~Describe blockLength here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(n, B, blockLength = 20)
{
        # Simple block bootstrap, overlapping blocks, no wrap-around,
        # no matching of ends.
        # n = length of data, B = bootstrap replicates
        nblocks <- ceiling(n / blockLength)
        x <- sample(1:(n - blockLength + 1), B * nblocks, replace = TRUE)
        dim(x) <- c(nblocks, B)
        apply(x, 2, function(y, L) (0:(L - 1)) + rep(y, each = L), L = blockLength)[1:n,]
  }

[Package openair version 1.0 Index]