smooth.trend(openair)	R Documentation

Calculate smooth trends

Description

Use non-parametric methods to calculate time series trends

Usage


smooth.trend(mydata,
pollutant = "nox",
deseason = FALSE,
type = "default",
simulate = FALSE,
n = 200,
autocor = FALSE,
ylab = pollutant,
main = "",
ci = FALSE,
alpha = 0.2,
auto.text = TRUE, ...)

Arguments

mydata 	 A data frame containing the field date and at least one other parameter for which a trend test is required; typically (but not necessarily) a pollutant.

pollutant 	 The parameter for which a trend test is required. Mandatory.

deseason 	 Should the data be de-deasonalized first? If TRUE the function stl is used (seasonal trend decomposition using loess). Note that if TRUE missing data are first linearly interpolated because stl cannot handle missing data.

type 	 The type of analysis to be done. The default is will produce a single plot using the entire data. Other types include hour (for hour of the day), weekday (for day of the week), month (for month of the year), wd (wind direction split into eight sectors and ws (wind speed split into eight quantiles). Finally, there is an option site, which requires a slightly different format. If site is used a data frame with headings date, a pollutant and site is required. It is also possible to choose type as another variable in the data frame. For example, type = "o3" will plot four wind roses for different levels of ozone, split into four quantiles (approximately equal numbers of counts in each of the four splits). This offers great flexibility for understanding the variation of different variables dependent on another.

simulate 	 Should simulations be carried out to determine the Mann-Kendall tau and p-value. The default is FALSE. If TRUE, bootstrap simulations are undertaken, which also account for autocorrelation.

n 	 Number of bootstrap simulations if simulate = TRUE.

autocor 	 Should autocorrelation be considered in the trend uncertainty estimates? The default is FALSE. Generally, accounting for autocorrelation increases the uncertainty of the trend estimate sometimes by a large amount.

ylab 	 y-axis label.

main 	 Title of plot, if required.

ci 	 Should confidence intervals be plotted? The default is FALSE.

alpha 	 The alpha transparency of shaded confidence intervals - if plotted. A value of 0 is fully transparent and 1 is fully opaque.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the ‘2’ in NO2.

... 	 Other graphical parameters e.g. pch = 16 for filled circles.

Details

The smooth.trend function provides a flexible way of estimating the trend in the concentration of a pollutant or other variable. Monthly mean values are calculated from an hourly (or higher resolution) or daily time series. There is the option to deseasonalise the data if there is evidence of a seasonal cycle.

smooth.trend uses a Generalized Additive Model (GAM) from the mgcv package to find the most appropriate level of smoothing. The function is particularly suited to situations where trends are not monotonic (see discussion with MannKendall for more details on this). The smooth.trend function is particularly useful as an exploratory technique e.g. to check how linear or non-linear trends are.

95% confidence intervals are shown by shading. Bootstrap estimates of the confidence intervals are also available through the simulate option. Residual resampling is used.

Trends can be considered in a very wide range of ways, controlled by setting type - see examples below.

Author(s)

David Carslaw

See Also

MannKendall for an alternative method of calculating trends.

Examples


# load example data from package
data(mydata)

# trend plot for nox
smooth.trend(mydata, pollutant = "nox")

# trend plot by each of 8 wind sectors
smooth.trend(mydata, pollutant = "o3", type = "wd", ylab = "o3 (ppb)")

[Package openair version 1.0 Index]