stl.plot(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


stl.plot(mydata, pollutant = "nox", s.wind = 17, ...)

Arguments

mydata 	 ~~Describe mydata here~~

pollutant 	 ~~Describe pollutant here~~

s.wind 	 ~~Describe s.wind here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(mydata, pollutant = "nox", s.wind = 17,...)  {
    #Does seasonal trend decompostion based on loess smoothing (STL)
    #NOTE! Currently this procedure will fill missing values by the mean of the whole
    # time series

    library(lattice)
    library(latticeExtra)
    #extract variables of interest
    vars <- c("date", pollutant)
    mydata <- mydata[, vars]
    mydata <- mydata[order(mydata$date), ]
    #rename pollutant to keep code consistent
    colnames(mydata)[2] <- "conc"

    means <- tapply(mydata$conc, format(mydata$date,"%Y-%m"), mean, na.rm = TRUE)

    means <- as.vector(means)

    #simple fill of missing data
    means[which(is.nan(means))] <- mean(means, na.rm = T)

    start.year <- as.numeric(format(mydata$date[1], "%Y"))
    end.year <- as.numeric(format(mydata$date[nrow(mydata)], "%Y"))
    start.month <- as.numeric(format(mydata$date[1], "%m"))
    end.month <- as.numeric(format(mydata$date[nrow(mydata)], "%m"))

    myts <- ts(means, start = c(start.year, start.month),
        end = c(end.year, end.month), frequency = 12)

    ssd <- stl(myts, s.window = s.wind, robust = TRUE, s.degree = 1)

    stlplot <- xyplot(ssd,
                xlab = "year",...,

                panel = function(x, y) {
                panel.abline(v = start.year:end.year, col = "grey85")
                panel.abline(h = 0, col = "grey85")
                panel.xyplot(x, y, type = "l")
                panel.xyplot(x, y, type = "p", cex = 0.2, col = "red")
                })

    print(stlplot)
    invisible(ssd)
  }

[Package openair version 1.0 Index]