summarise(openair)	R Documentation

Function to rapidly provide an overview of air quality data

Description

This function provides a quick graphical and numerical summary of data. The location presence/absence of data are shown, with summary statistics and plots of variable distribitions.

Usage


summarise(mydata,
na.len = 24,
clip = TRUE,
percentile = 0.99,
type = "histogram",
breaks = 20,
col.data = "lightblue",
col.mis = "red",
col.hist = "forestgreen",
auto.text = TRUE)

Arguments

mydata 	 A data frame to be summarised. Must contain a date field and at least one other parameter.

na.len 	 Missing data are only shown with at least na.len contiguous missing vales. The puprose of setting na.len is for clarity: with long time series it is difficult to see where individual missing hours are. Furthermore, setting na.len = 96, for example would show where there are at least 4 days of continuous missing data.

clip 	 When data contain outliers, the histogram or density plot can fail to show the distribution of the main body of data. Setting clip = TRUE, will remove the top 1 % of data to yield what is often a better display of the overall distribution of the data. The amount of clipping can be set with percentile.

percentile 	 This is used to clip the data. For example, percentile = 0.99 (the default) will remove the top 1 percentile of values i.e. values greater than the 99th percentile will not be used.

type 	 type is used to determine whether a histogram (the default) or a denisty plot is used to show the distribution of the data.

breaks 	 Number of histogram bins. Sometime useful but not easy to set a single value for a range of very different variables.

col.data 	 Colour to be used to show the presence of data. Type colors() into R to see the full range of colour names.

col.mis 	 Colour to be used to show missing data.

col.hist 	 Colour for the histogram or desnity plot.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the ‘2’ in NO2.

Details

summarise produces two panels of plots: one showing the presence/absence of data and the other the distributions. The left panel shows time series and codes the presence or adsence of data in different colours. By stacking the plots one on top of another it is easy to compare different pollutants/variables. Overall statistics are given for each variable: mean, maximum, minimum, missing hours (also expressed as a percentage), median and the 95th percentile. For each year the data capture rate (expressed as a percentage of hours in that year) is also given.

The right panel shows either a histogram or a density plot depending on the choice of type. Density plots avoid the issue of arbitrary bin sizes that can sometimes provide a misleading view of the data distribution. Density plots are often more appropriate, but their effectiveness will depend on the data in question.

Author(s)

David Carslaw

Examples



# load example data from package
data(mydata)

# do not clip density plot data
summarise(mydata, clip = FALSE)

# exclude highest 5 % of data etc.
summarise(mydata, percentile = 0.95)

# show missing data where there are at least 10 contiguous missing values
summarise(mydata, na.len = 10)

# show data in green
summarise(mydata, col.data = "green")

# show missing data in yellow
summarise(mydata, col.mis = "yellow")

# show density plot line in black
summarise(mydata, col.dens = "black")


[Package openair version 1.0 Index]