tau(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


tau(x, y, alpha = 0.05)

Arguments

x 	 ~~Describe x here~~

y 	 ~~Describe y here~~

alpha 	 ~~Describe alpha here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(x, y, alpha = 0.05){
#
#   Compute Kendall's tau plus a 1-alpha confidence interval
#   using the method recommended by Long and Cliff (1997).
#
        xdif <- outer(x, x, FUN = "-")
        ydif <- outer(y, y, FUN = "-")
        tv <- sign(xdif) * sign(ydif)
        dbar <- apply(tv, 1, mean)
        n <- length(x)
        tau <- sum(tv) / (n * (n - 1))
        A <- sum((dbar - tau)^2) / (n - 1)
        B <- (n * (n - 1) * (-1) * tau^2 + sum(tv^2)) / (n^2 - n - 1)
        C <- (4 * (n - 2) * A + 2 * B)/(n * (n - 1))
        crit <- qnorm(alpha / 2)
        cilow <- tau + crit * sqrt(C)
        cihi <- tau - crit * sqrt(C)
        test <- tau / sqrt((2 * (2 * n + 5))/(9 * n * (n - 1)))
        siglevel <- 2 * (1 - pnorm(abs(test)))
        list(cor = tau, ci = c(cilow, cihi), siglevel = siglevel)
  }

[Package openair version 1.0 Index]