tau.boot(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


tau.boot(y)

Arguments

y 	 ~~Describe y here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(y){
#
#   Compute Kendall's tau plus a 1-alpha confidence interval
#   using the method recommended by Long and Cliff (1997).
#
        x <- 1: length(y)
        xdif <- outer(x, x, FUN = "-")
        ydif <- outer(y, y, FUN = "-")
        tv <- sign(xdif) * sign(ydif)
        dbar <- apply(tv, 1, mean)
        n <- length(x)
        tau <- sum(tv) / (n * (n - 1))
        
        list(cor = tau)
  }

[Package openair version 1.0 Index]