trend.hour.weekday(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


trend.hour.weekday(mydata, pollutant = "nox", limits = c(0, 100), line.colour = "darkorange2", fill.colour = "lightgoldenrod", ylab = pollutant, main = "", auto.text = TRUE, ...)

Arguments

mydata 	 ~~Describe mydata here~~

pollutant 	 ~~Describe pollutant here~~

limits 	 ~~Describe limits here~~

line.colour 	 ~~Describe line.colour here~~

fill.colour 	 ~~Describe fill.colour here~~

ylab 	 ~~Describe ylab here~~

main 	 ~~Describe main here~~

auto.text 	 ~~Describe auto.text here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(mydata,
                pollutant = "nox",
                limits = c(0, 100),
                line.colour = "darkorange2",
                fill.colour = "lightgoldenrod",
                                ylab = pollutant,
                                main = "",
                                auto.text = TRUE,...) {
    library(lattice)
    library(latticeExtra)
    library(Hmisc)

trend.hour.weekday.plot <- function(mydata, pollutant, limits, line.colour,
                fill.colour,...) {

    #extract variables of interest
    vars <- c("date", pollutant)
    mydata <- mydata[, vars]
    mydata <- mydata[order(mydata$date), ]
    
    colnames(mydata)[2] <- "conc"

    year <- as.factor(format(mydata$date,"%Y"))
    hour <- as.numeric(format(mydata$date,"%H"))
    weekday <- format(mydata$date, "%A")
    
    mydata <- cbind(mydata, year, weekday, hour)

    #summarize data
    s <- summarize(mydata$conc, llist(year, weekday, hour), smean.cl.normal)
    
    s$weekday <- ordered(s$weekday, levels = c("Monday", "Tuesday",
                  "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"))

  xyplot(mydata.conc ~ hour | year + weekday,
      data = s,
      col = line.colour,
      type = "l",
      as.table = TRUE,
          main = quick.text(main, auto.text),
          ylab = quick.text(ylab, auto.text),
      scales = list(x = list(at = seq(0, 24, 6))),
      xlab = "hour of day",...,

      panel = function(x, y, subscripts,...) {

      #for polygon
        x1 <- s$hour[subscripts]
        x2 <- x1
        y1 <- s$Lower[subscripts]
        y2 <- s$Upper[subscripts]

        #clever polygon that can handle missing data
          for(i in seq(2, length(x1)))

          if (!any(is.na(y2[c(i - 1, i)]))){
            lpolygon(c(x1[i - 1], x1[i], x2[i], x2[i-1]),
                c(y1[i - 1], y1[i], y2[i], y2[i-1]),
                col = fill.colour, border = 0) }

              panel.grid(h = -1, v = 0, lwd = 0.5, col = "grey85")
              panel.abline(v = seq(0, 24, 6), col = "grey85")
              panel.xyplot(x, y, lwd = 2,...)
      }
)}

plt <- trend.hour.weekday.plot(mydata, pollutant, limits, line.colour,
                fill.colour,...)

    useOuterStrips(plt, strip = strip.custom(par.strip.text = list(cex = 0.8)),
    strip.left = strip.custom(par.strip.text = list(cex = 0.8)))
  }

[Package openair version 1.0 Index]