trend.plot(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


trend.plot(mydata, pollutant = "nox", deseason = FALSE, simulate = FALSE, n = 200, legend.pos = "bottomleft", ...)

Arguments

mydata 	 ~~Describe mydata here~~

pollutant 	 ~~Describe pollutant here~~

deseason 	 ~~Describe deseason here~~

simulate 	 ~~Describe simulate here~~

n 	 ~~Describe n here~~

legend.pos 	 ~~Describe legend.pos here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(mydata, 
                      pollutant = "nox", 
                      deseason = FALSE, 
                      simulate = FALSE,
                      n = 200,
                      legend.pos = "bottomleft",...)  {        #bootstrap simulations

library(mgcv)    
    #extract variables of interest
    vars <- c("date", pollutant)
    mydata <- mydata[, vars]
    mydata <- mydata[order(mydata$date), ]
    
    #rename pollutant to keep code consistent
    colnames(mydata)[2] <- "conc"

    #sometimes data have long trailing NAs, so start and end at first and last data
    min.idx <- min(which(!is.na(mydata$conc)))
    max.idx <- max(which(!is.na(mydata$conc)))
    
    mydata <- mydata[min.idx:max.idx, ]

    means <- tapply(mydata$conc, format(mydata$date,"%Y-%m"), mean, na.rm = TRUE)
    means <- as.vector(means)
    
    start.year <- as.numeric(format(mydata$date[1], "%Y"))
    end.year <- as.numeric(format(mydata$date[nrow(mydata)], "%Y"))
    start.month <- as.numeric(format(mydata$date[1], "%m"))
    end.month <- as.numeric(format(mydata$date[nrow(mydata)], "%m"))

    #dates as numeric for GAM fitting
    dates <- seq((start.year + (start.month - 1)/12), 
             (end.year + (end.month - 1)/12), by = 1/12)
    
    if (deseason == TRUE) {
    
        #simple fill of missing data
        means[which(is.nan(means))] <- mean(means, na.rm = TRUE)
        myts <- ts(means, start = c(start.year, start.month),
                end = c(end.year, end.month), frequency = 12)
    
        ssd <- stl(myts, "periodic", robust = TRUE, s.degree = 1)
        
        deseas <- ssd$time.series[, "trend"] + ssd$time.series[, "remainder"]
        results <- data.frame(date = dates, conc = deseas)
        
    } else {
    
        results <- na.omit(data.frame(date = dates, conc = means))
    
    }
    
    if (simulate == FALSE) {  #just plot the data   
        mod <- gam(conc ~ s(date), data = results)
        pred <- predict(mod, results, se = TRUE)
        
        results <- cbind(results, pred = pred$fit, lower = pred$fit - 2 * pred$se.fit, 
                    upper = pred$fit + 2 * pred$se.fit)
        
        x <- c(results$date, rev(results$date))
        y <- c(pred$fit + 2 * pred$se.fit, rev(pred$fit - 2 * pred$se.fit))
        
        plot(results$date, results$conc, type = "n", xlab = "date",...)
        
        polygon(x, y, col = "lightpink", border = NA)
        
        grid(col = "grey85", lty = 1)      
        lines(results$date, pred$fit, col = "red", lwd = 2)
        lines(results$date, results$conc, col = "grey50")
    } else {
   
       sam.size = nrow(results)
       boot.pred <- matrix(nrow = sam.size, ncol = n)
       boot.pred2 <- matrix(nrow = sam.size, ncol = n)
       
       for (i in 1:n) {
          index <- sample(sam.size, replace = TRUE)
          index2 <- sample(sam.size, replace = TRUE)
          
          new.results <- data.frame(date = results$date[index], 
                         conc = results$conc[index2])
                        
          mod <- gam(conc ~ s(date, bs = "tp"), data = results[index, ])
          mod2 <- gam(conc ~ s(date, bs = "tp"), data = new.results)
          
          pred <- predict(mod, results)
          pred2 <- predict(mod2, new.results) 
          
          boot.pred[, i] <- pred
          boot.pred2[, i] <- pred2
       }   
       
       #calculate percentiles
       percentiles <- apply(boot.pred, 1, function(x) 
                      quantile(x, probs = c(0.025, 0.975)))
       percentiles2 <- apply(boot.pred2, 1, function(x) 
                       quantile(x, probs = c(0.025, 0.975))) 
       
       x <- c(results$date, rev(results$date))
       y <- c(percentiles[2, ], rev(percentiles[1, ]))
          
        plot(results$date, results$conc, type = "n", xlab = "date",...)
        
        grid(col = "grey85", lty = 1)
        lines(results$date, results$conc, col = "grey50")
        polygon(x, y, col = rgb(1, 0, 0, 0.2), border = NA)
        
        y <- c(percentiles2[2, ], rev(percentiles2[1, ]))
        polygon(x, y, col = rgb(0, 0, 1, 0.2), border = NA)
              
        lines(results$date, rowMeans(boot.pred), col = "red", lwd = 2)
        lines(results$date, rowMeans(boot.pred2), col = "blue", lwd = 2)
        
       legend(legend.pos, c("Measurements", "No trend hypothesis"), col = c("red",
              "blue"), lty = 1, lwd = 2, inset = 0.02, bg = "white")
    }
   results
   invisible(results)    
  }

[Package openair version 1.0 Index]