tsp1reg(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


tsp1reg(x, y, plotit = FALSE)

Arguments

x 	 ~~Describe x here~~

y 	 ~~Describe y here~~

plotit 	 ~~Describe plotit here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(x, y, plotit = FALSE){
#
# Compute the Theil-Sen regression estimator.
# Only a single predictor is allowed in this version
#
        temp <- matrix(c(x,y), ncol = 2)
        temp <- elimna(temp)     # Remove any pairs with missing values
        x <- temp[,1]
        y <- temp[,2]
        ord <- order(x)
        xs <- x[ord]
        ys <- y[ord]
        vec1 <- outer(ys, ys, "-")
        vec2 <- outer(xs, xs, "-")
        v1 <- vec1[vec2 > 0]
        v2 <- vec2[vec2 > 0]
        slope <- median(v1 / v2)
        coef <- median(y) - slope * median(x)
        names(coef) <- "Intercept"
        coef <-c(coef, slope)
        if(plotit){
                plot(x, y, xlab="X", ylab="Y")
                abline(coef)
        }
        res <- y - slope * x - coef[1]
        list(coef = coef, residuals = res)
  }

[Package openair version 1.0 Index]