wind.rose(openair)	R Documentation

Traditional wind rose plot

Description

Traditional wind rose plot that plots wind speed and wind direction by different intervals

Usage


wind.rose(polar,
ws.int = 2,
angle = 30,
type = "default",
cols = "default",
main = "",
grid.line = 5,
width = 1,
units = "(m/s)",
auto.text = TRUE,
 ...)

Arguments

polar 	 A data frame containing fields ws and wd

ws.int 	 Wind speed interval. Default is 2 m/s but for low met masts with low mean wind speeds a value of 1 or 0.5 m/s may be better.

angle 	 Default angle of "spokes" is 30. Other potentially useful angles are 45 and 10. Note that the width of the wind speed interval may need adjusting using width.

type 	 Default is default, which will plot all data on one wind rose. Can also be year, month and hour for one wind rose for each year or each month respectfully.

cols 	 Colours to be used for plotting. Options include default, increment, heat, jet, hue and user defined. For user defined the user can supply a list of colour names recognised by R (type colours() to see the full list). An example would be cols = c("yellow", "green", "blue", "black")

main 	 Title of plot.

grid.line 	 Grid line interval to use.

width 	 Adjustment factor for width of wind speed intervals. For example, width = 1.5 will make the width 1.5 times wider.

units 	 Controls the units shown on the scale e.g. units = "knots".

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the `2' in NO2.

... 	 Other graphical parameters to pass to wind.rose.

Details

Data are summarised by direction, typically by 45 or 30 (or 10) degrees and by different wind speed categories. Typically, wind speeds are represented by different width "paddles". The plots show the proportion (here represented as a percentage) of time that the wind is from a certain angle and wind speed range.

The wind.rose function plots the frequency of wind directions/speeds in 30 degree intervals of wind direction by default. By default the function will show a wind rose using all the data. Sometimes it is useful to set the wind speed interval to < 2 m/s if wind speeds are from a met mast that is lower than 10 m height.

Author(s)

David Carslaw

See Also

polar.freq for a more flexible version that considers other statistics and pollutant concentrations.

Examples


# load example data from package data(mydata)

# basic plot
wind.rose(mydata)

# one wind rose for each year
wind.rose(mydata,type = "year")

# plot wind roses by month of the year
wind.rose(mydata, type = "month")

# wind rose in 10 degree intervals with gridlines and width adjusted
wind.rose(mydata, angle = 10, width = 0.3, grid.line = 1)

[Package openair version 1.0 Index]