year.panels(openair)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


year.panels(mydata, pollutant = "nox", add.smooth = FALSE, ylab = pollutant, main = "", auto.text = TRUE, ...)

Arguments

mydata 	 ~~Describe mydata here~~

pollutant 	 ~~Describe pollutant here~~

add.smooth 	 ~~Describe add.smooth here~~

ylab 	 ~~Describe ylab here~~

main 	 ~~Describe main here~~

auto.text 	 ~~Describe auto.text here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(mydata,
                pollutant = "nox",
                add.smooth = FALSE,
                                ylab = pollutant,
                                main = "",
                                auto.text = TRUE,...) {

    library(lattice)

    #extract variables of interest
    vars <- c("date", pollutant)
    mydata <- mydata[, vars]
    mydata <- mydata[order(mydata$date), ]
    #rename pollutant to keep code consistent
    colnames(mydata)[2] <- "conc"
    
    year <- as.factor(format(mydata$date, "%Y"))
    mydata <- cbind(mydata, year)

    #determine begin/end year (+1) for gridlines and axis
    begin.year <- ISOdate(levels(year)[1], 1, 1, 0, 0)
    end.year <- ISOdate(as.numeric(levels(year)[length(levels(year))]) +1, 1, 1, 0, 0)

    xyplot(conc ~ date | year,
        data = mydata,
        aspect = 0.4,
        as.table = TRUE,
                main = quick.text(main, auto.text),
                ylab = quick.text(ylab, auto.text),
        scales = list(relation = "free",x = list(format = "%b",
        at = seq(begin.year, end.year, by = "2 month"))),...,

        panel = function(x, y,...) {
          #add grid lines every month by finding start/end date
          panel.abline(v = seq(begin.year, end.year, by = "month"), col = "grey85")
          panel.abline(h = 0, col = "grey85")
          panel.grid(h = -1, v = 0)
          panel.xyplot(x, y, type = "l", lwd = 1,...)

          if (add.smooth == TRUE) {
              panel.loess(x, y, lwd = 2, col = "red")
          }
    })
  }

[Package openair version 1.0 Index]