Import pre-calculated airbase statistics

Import pre-calculated airbase statistics

airbaseStats(statistic = "Mean", pollutant = "no2", avg.time = "auto",
  code = NULL, site.type = c("background", "traffic", "industrial",
  "unknown"), year = 1969:2012, data.cap = 0, add = c("country", "lat",
  "lon", "site.type", "site", "city"))

Arguments

statistic
A single choice from “P50”, “Mean”, “P95”, “P98”, “Max”, “Max36”, “Max8”, “Days.c.50.”, “Max26”, “Days.c.120.”,“SOMO35”, “AOT40”, “Max19”, “Hours.c.200.”, “Max4”, “Days.c.125.”, “P99_9”, “Max25”, “Hours.c.350.”.

pollutant
The single name of a pollutant to extract; can be upper or lower case but must match one of those below.

avg.time
The averaging time on which the measurements are based. By default the averaging period with most data associated with it is used. For most pollutants the averaging period will be “day”. Other common options are “hour” and “week”.

code
The airbase site code(s) to select. Can be upper or lower case.

site.type
The type of site(s) to select.

year
The year(s) to select e.g. year = 2000:2012 to select years 2010, 2011 and 2012.

data.cap
The data capture threshold to use (%). By default all data are selected.

add
Additional fields to add to the returned data frame. By default the country, site type, latitude and logitude are returned. Other useful options include “city”, “site” (site name), “EMEP_station” and “altitude”.

Value

A data frame of airbase sites with the statsitics chosen and for all species.

Returns a data frame with POSIXct date, together with many other airbase fields.

Details

The airbase database makes available detailed pre-calculated statsitics for all species, sites and years. The airbaseStats function provides these data.

The data are imported depending on the statistic chosen --- see the list of available statistics in the statistics option. Further options are available to refine the selection including a single pollutant, year(s), site type(s), data capture percentage threshold and airbase site code.

Once the data have been imported it is then possible to subset the data in various ways e.g. by extracting a particular species with a defined data capture threshold.

The following pollutants are available:

(CH3)2-CH-CH2-CH2-CH3

(CH3)3-C-CH2-CH-(CH3)2

1,2,3-C6H3(CH3)3

1,2,4-C6H3(CH3)3

1,3,5-C6H3(CH3)3

3-methylpentane

Acenaphthene

Acenaphtylene

Al

Anthracene

As

As in PM10

As in PM2.5

As in TSP

BS

BaP

BaP in PM10

BaP in PM2.5

Benzo(a)anthracene

Benzo(a)anthracene in PM10

Benzo(b)fluoranthene

Benzo(b)fluoranthene in PM10

Benzo(b+j+k)fluoranthenes

Benzo(b,j,k)fluoranthene

Benzo(b,j,k)fluorantheneInPM1

Benzo(e)pyrene

Benzo(ghi)perylene

Benzo(j)fluoranthene

Benzo(j)fluoranthene in PM10

Benzo(k)fluoranthene

Benzo(k)fluoranthene in PM10

Black Carbon

C2Cl4

C2H6

C6H14

C6H5-C2H5

C6H5-CH3

C6H5-CH=CH2

C6H6

C7H16

C8H18

CH2=CH-C(CH3)=CH2

CH2=CH-CH3

CH2=CH-CH=CH2

CH4

CHCl=CCl2

CO

CO2

CS2

Ca++

Ca2+ in PM10

Ca2+ in PM2.5

Cd

Cd in PM10

Cd in PM2.5

Cd in TSP

Chrysene

Chrysene + Triphenylene

Cl-

Cl- in PM10

Cl- in PM2.5

Co

Cr

Cu

Cyclo-hexane

Dibenzo(ah)anthracene

Dibenzo(ah)anthracene in PM10

EC in PM10

EC in PM2.5

Fe

Fluoranthene

Fluorene

H+

H2C=CH-CH2-CH2-CH3

H2C=CH-CH2-CH3

H2C=CH2

H2S

H3C-(CH2)3-CH3

H3C-CH(CH3)2

H3C-CH2-CH(CH3)2

H3C-CH2-CH2-CH3

H3C-CH2-CH3

H3C-HC=CH-CH2-CH3

HC C2-C6(excl. AROM. & CHLH)

HC=CH

HCHO

HCl

HF

HNO3+NO3

Hg

Hg in PM10

Hg-reactive

Hg0

Hg0 + Hg-reactive

Indeno-(1,2,3-cd)pyrene

Indeno-(1,2,3-cd)pyrene in PM

K+

K+ in PM10

K+ in PM2.5

Mg++

Mg2+ in PM10

Mg2+ in PM2.5

Mn

N-DEP

N2O

NH3

NH3+NH4

NH4

NH4+

NH4+ in PM10

NH4+ in PM2.5

NO

NO2

NO3

NO3-

NO3- in PM10

NO3- in PM2.5

NOX

Na+

Na+ in PM10

Na+ in PM2.5

Naphtalene

Ni

Ni in PM10

Ni in PM2.5

Ni in TSP

O3

OC in PM10

OC in PM2.5

PAH

PAN

PCB-105

PCB-114

PCB-118

PCB-138

PCB-141

PCB-149

PCB-153

PCB-156

PCB-157

PCB-167

PCB-170

PCB-18

PCB-180

PCB-183

PCB-187

PCB-189

PCB-194

PCB-28

PCB-31

PCB-52

PCB-74

PCB-99

PCB_123

PM1

PM10

PM2.5

Pb

Pb in PM10

Pb in PM2.5

Pb in TSP

Phenanthrene

Pyrene

Radioactivity

S-DEP

SA

SO2

SO2 + SO4--

SO4 (H2SO4 aerosols) (SO4--)

SO4--

SO42- in PM10

SO42- in PM2.5

SPM

Se

T-VOC

THC (NM)

V

Vanadium

Zn

Zn in PM2.5

cis-H3C-CH=CH-CH3

k

m,p-C6H4(CH3)2

o-C6H4-(CH3)2

pH

precip_amount

precip_amount_off

trans-H3C-CH=CH-CH3