Import air quality data from European database

This function is a simplified version of the saqgetr package (see https://github.com/skgrange/saqgetr) for accessing European air quality data. The function only returns valid hourly data and is meant as a fast and convenient way of accessing the most common type of hourly air quality data. The function works in the same way as other openair functions that import air quality data that generally need a site code and year to be supplied.

Usage

importEurope(
  site = "debw118",
  year = 2018,
  tz = "UTC",
  meta = FALSE,
  to_narrow = FALSE,
  progress = TRUE
)

Arguments

site

The code of the site(s).

year

Year or years to import. To import a sequence of years from 1990 to 2000 use year = 1990:2000. To import several specific years use year = c(1990, 1995, 2000) for example.

tz

Not used

meta

Should meta data be returned? If TRUE the site type, latitude and longitude are returned.

to_narrow

By default the returned data has a column for each pollutant/variable. When to_narrow = TRUE the data are stacked into a narrow format with a column identifying the pollutant name.

progress

Show a progress bar when many sites/years are being imported? Defaults to TRUE.

Value

a tibble

Details

The function can however return key site meta data.

The saqgetr package is much more comprehensive and provides data at other time averages e.g. daily data.

See also

Examples

# import data for Stuttgart Am Neckartor (S)
if (FALSE) { # \dontrun{
stuttgart <- importEurope("debw118", year = 2010:2019, meta = TRUE)
} # }