Import data from King's College London networks

Function for importing hourly mean data from King's College London networks. Files are imported from a remote server operated by King's College London that provides air quality data files as R data objects.

importKCL(
  site = "my1",
  year = 2009,
  pollutant = "all",
  met = FALSE,
  units = "mass",
  extra = FALSE,
  meta = FALSE,
  to_narrow = FALSE
)

Arguments

site
Site code of the network site to import e.g. "my1" is Marylebone Road. Several sites can be imported with site = c("my1", "kc1") --- to import Marylebone Road and North Kensignton for example.

year
Year or years to import. To import a sequence of years from 1990 to 2000 use year = 1990:2000. To import several specfic years use year = c(1990, 1995, 2000) for example.

pollutant
Pollutants to import. If omitted will import all pollutants from a site. To import only NOx and NO2 for example use pollutant = c("nox", "no2").

met
Should meteorological data be added to the import data? The default is FALSE. If TRUE wind speed (m/s), wind direction (degrees), solar radiation and rain amount are available. See details below.

Access to reliable and free meteorological data is problematic.

units
By default the returned data frame expresses the units in mass terms (ug/m3 for NOx, NO2, O3, SO2; mg/m3 for CO). Use units = "volume" to use ppb etc. PM10_raw TEOM data are multiplied by 1.3 and PM2.5 have no correction applied. See details below concerning PM10 concentrations.

extra
Not currently used.

meta
Should meta data be returned? If TRUE the site type, latitude and longitude are returned.

to_narrow
By default the returned data has a column for each pollutant/variable. When to_narrow = TRUE the data are stacked into a narrow format with a column identifying the pollutant name.

Value

Returns a data frame of hourly mean values with date in POSIXct class and time zone GMT.

Details

The importKCL function has been written to make it easy to import data from the King's College London air pollution networks. KCL have provided .RData files (R workspaces) of all individual sites and years for the KCL networks. These files are updated on a weekly basis. This approach requires a link to the Internet to work.

There are several advantages over the web portal approach where .csv files are downloaded. First, it is quick to select a range of sites, pollutants and periods (see examples below). Second, storing the data as .RData objects is very efficient as they are about four times smaller than .csv files --- which means the data downloads quickly and saves bandwidth. Third, the function completely avoids any need for data manipulation or setting time formats, time zones etc. Finally, it is easy to import many years of data beyond the current limit of about 64,000 lines. The final point makes it possible to download several long time series in one go. The function also has the advantage that the proper site name is imported and used in openair functions.

The site codes and pollutant names can be upper or lower case. The function will issue a warning when data less than six months old is downloaded, which may not be ratified.

The data are imported by stacking sites on top of one another and will have field names date, site, code (the site code) and pollutant(s). Sometimes it is useful to have columns of site data. This can be done using the reshape function --- see examples below.

The situation for particle measurements is not straightforward given the variety of methods used to measure particle mass and changes in their use over time. The importKCL function imports two measures of PM10 where available. PM10_raw are TEOM measurements with a 1.3 factor applied to take account of volatile losses. The PM10 data is a current best estimate of a gravimetric equivalent measure as described below. NOTE! many sites have several instruments that measure PM10 or PM2.5. In the case of FDMS measurements, these are given as separate site codes (see below). For example "MY1" will be TEOM with VCM applied and "MY7" is the FDMS data.

Where FDMS data are used the volatile and non-volatile components are separately reported i.e. v10 = volatile PM10, v2.5 = volatile PM2.5, nv10 = non-volatile PM10 and nv2.5 = non-volatile PM2.5. Therefore, PM10 = v10 + nv10 and PM2.5 = v2.5 + nv2.5.

For the assessment of the EU Limit Values, PM10 needs to be measured using the reference method or one shown to be equivalent to the reference method. Defra carried out extensive trials between 2004 and 2006 to establish which types of particulate analysers in use in the UK were equivalent. These trials found that measurements made using Partisol, FDMS, BAM and SM200 instruments were shown to be equivalent to the PM10 reference method. However, correction factors need to be applied to measurements from the SM200 and BAM instruments. Importantly, the TEOM was demonstrated as not being equivalent to the reference method due to the loss of volatile PM, even when the 1.3 correction factor was applied. The Volatile Correction Model (VCM) was developed for Defra at King's to allow measurements of PM10 from TEOM instruments to be converted to reference equivalent; it uses the measurements of volatile PM made using nearby FDMS instruments to correct the measurements made by the TEOM. It passed the equivalence testing using the same methodology used in the Defra trials and is now the recommended method for correcting TEOM measurements (Defra, 2009). VCM correction of TEOM measurements can only be applied after 1st January 2004, when sufficiently widespread measurements of volatile PM became available. The 1.3 correction factor is now considered redundant for measurements of PM10 made after 1st January 2004. Further information on the VCM can be found at http://www.volatile-correction-model.info/.

All PM10 statistics on the LondonAir web site, including the bulletins and statistical tools (and in the RData objects downloaded using importKCL), now report PM10 results as reference equivalent. For PM10 measurements made by BAM and SM200 analysers the applicable correction factors have been applied. For measurements from TEOM analysers the 1.3 factor has been applied up to 1st January 2004, then the VCM method has been used to convert to reference equivalent.

The meteorological data are meant to represent 'typical' conditions in London, but users may prefer to use their own data. The data provide a an estimate of general meteorological conditions across Greater London. For meteorological species (wd, ws, rain, solar) each data point is formed by averaging measurements from a subset of LAQN monitoring sites that have been identified as having minimal disruption from local obstacles and a long term reliable dataset. The exact sites used varies between species, but include between two and five sites per species. Therefore, the data should represent 'London scale' meteorology, rather than local conditions.

While the function is being developed, the following site codes should help with selection. We will also make available other meta data such as site type and location to make it easier to select sites based on other information. Note that these codes need to be refined because only the common species are available for export currently i.e. NOx, NO2, O3, CO, SO2, PM10, PM2.5.

A30 | Kingston - Kingston Bypass A3 | Roadside

AD1 | Shoreham-by-Sea | Kerbside

AR1 | Chichester - Lodsworth | Rural

AR2 | Wealden - Isfield | Rural

AS1 | Bath Aethalometer | Urban Background

BA1 | Basildon - Gloucester Park | Roadside

BB1 | Broxbourne (Roadside) | Roadside

BE0 | Belfast - Carbon | Urban Background

BE1 | Belfast Centre AURN | Urban Background

BE3 | Belfast Centre Aethalometer | Urban Background

BE7 | Belfast Centre FDMS trial | Urban Background

BE8 | Belfast - Nitrate | Urban Background

BE9 | Belfast - Partisol SO4 | Urban Background

BF1 | Bedford Stewartby (Rural) | Industrial

BF3 | Bedford - Kempston | Industrial

BF4 | Bedford - Prebend Street | Roadside

BF5 | Bedford - Lurke Street | Roadside

BG1 | Barking and Dagenham - Rush Green | Suburban

BG2 | Barking and Dagenham - Scrattons Farm | Suburban

BG3 | Barking and Dagenham - North Street | Kerbside

BH0 | Brighton Preston Park AURN | Urban Background

BH1 | Brighton Roadside | Roadside

BH2 | Brighton and Hove - Hove Town Hall | Roadside

BH3 | Brighton and Hove - Foredown Tower | Urban Background

BH5 | Brighton Mobile (Preston Fire Station) | Roadside

BH6 | Brighton Mobile (Lewes Road) | Roadside

BH7 | Brighton Mobile (Gloucester Road) | Roadside

BH8 | Brighton and Hove - Stanmer Park | Rural

BH9 | Brighton Mobile Beaconsfield Road | Roadside

BI1 | Birmingham Tyburn CPC | Urban Background

BL0 | Camden - Bloomsbury | Urban Background

BL1 | Bloomsbury AURN SMPS | Urban Background

BM1 | Ballymena - Ballykeel | Suburban

BM2 | Ballymena - North Road | Roadside

BN1 | Barnet - Tally Ho Corner | Kerbside

BN2 | Barnet - Finchley | Urban Background

BN3 | Barnet - Strawberry Vale | Urban Background

BO1 | Ballymoney 1 | Suburban

BP0 | Westminster - Bridge Place | Urban Background

BQ5 | Bexley - Manor Road West Gravimetric | Industrial

BQ6 | Bexley - Manor Road East Gravimetric | Industrial

BQ7 | Belvedere West | Urban Background

BQ8 | Belvedere West FDMS | Urban Background

BT1 | Brent - Kingsbury | Suburban

BT2 | Brent - Ikea Car Park | Roadside

BT3 | Brent - Harlesden | Roadside

BT4 | Brent - Ikea | Roadside

BT5 | Brent - Neasden Lane | Industrial

BT6 | Brent - John Keble Primary School | Roadside

BT7 | Brent - St Marys Primary School | Urban Background

BW1 | Brentwood - Brentwood Town Hall | Urban Background

BX0 | Bexley - Belvedere FDMS | Suburban

BX1 | Bexley - Slade Green | Suburban

BX2 | Bexley - Belvedere | Suburban

BX3 | Bexley - Thamesmead | Suburban

BX4 | Bexley - Erith | Industrial

BX5 | Bexley - Bedonwell | Suburban

BX6 | Bexley - Thames Road North FDMS | Roadside

BX7 | Bexley - Thames Road North | Roadside

BX8 | Bexley - Thames Road South | Roadside

BX9 | Bexley - Slade Green FDMS | Suburban

BY1 | Bromley - Rent Office | Urban Background

BY4 | Bromley - Tweedy Rd | Roadside

BY5 | Bromley - Biggin Hill | Suburban

BY7 | Bromley - Harwood Avenue | Roadside

CA1 | Crawley Background | Urban Background

CA2 | Crawley - Gatwick Airport | Urban Background

CB1 | Chelmsford - Fire Station | Roadside

CB2 | Chelmsford - Springfield Road | Roadside

CB3 | Chelmsford - Chignal St James | Urban Background

CB4 | Chelmsford - Baddow Road | Roadside

CC1 | Colchester - Lucy Lane South | Roadside

CC2 | Colchester - Brook Street | Roadside

CC3 | Colchester - Mersea Road | Roadside

CD1 | Camden - Swiss Cottage | Kerbside

CD3 | Camden - Shaftesbury Avenue | Roadside

CD4 | Camden - St Martins College (NOX 1) | Urban Background

CD5 | Camden - St Martins College (NOX 2) | Urban Background

CD7 | Camden - Swiss Cottage Partisol | Kerbside

CD9 | Camden - Euston Road | Roadside

CF1 | Cardiff Aethalometer | Urban Background

CH1 | Cheltenham | Urban Background

CI1 | Chichester - A27 Chichester Bypass | Roadside

CI4 | Chichester - Orchard Street | Roadside

CK1 | Cookstown | Suburban

CP1 | Castle Point - Canvey Island | Urban Background

CR2 | Croydon - Purley Way | Roadside

CR3 | Croydon - Thornton Heath | Suburban

CR4 | Croydon - George Street | Roadside

CR5 | Croydon - Norbury | Kerbside

CR6 | Croydon - Euston Road | Suburban

CT1 | City of London - Senator House | Urban Background

CT2 | City of London - Farringdon Street | Kerbside

CT3 | City of London - Sir John Cass School | Urban Background

CT4 | City of London - Beech Street | Roadside

CT6 | City of London - Walbrook Wharf | Roadside

CT8 | City of London - Upper Thames Street | Roadside

CY1 | Crystal Palace - Crystal Palace Parade | Roadside

DC1 | Dacorum 1 Hemel Hempstead (Background) | Urban Background

DC2 | Dacorum 2 Hemel Hempstead (Background) | Urban Background

DC3 | High Street Northchurch | Roadside

DE1 | Derry City - Brandywell | Urban Background

DE2 | Derry City - Dales Corner | Roadside

DM1 | Dunmurry Aethalometer | Urban Background

EA0 | Ealing - Acton Town Hall FDMS | Roadside

EA1 | Ealing - Ealing Town Hall | Urban Background

EA2 | Ealing - Acton Town Hall | Roadside

EA3 | Ealing 3 - A40 East Acton | Roadside

EA4 | Ealing Mobile - Hamilton Road | Roadside

EA5 | Ealing Mobile - Southall | Roadside

EA6 | Ealing - Hanger Lane Gyratory | Roadside

EA7 | Ealing - Southall | Urban Background

EA8 | Ealing - Horn Lane | Industrial

EA9 | Ealing - Court Way | Roadside

EB1 | Eastbourne - Devonshire Park | Urban Background

EB3 | Eastbourne - Holly Place | Urban Background

EH1 | E Herts Throcking (Rural) | Rural

EH2 | East Herts Sawbridgeworth (Background) | Urban Background

EH3 | East Herts Sawbridgeworth (Roadside) | Roadside

EH4 | East Herts Ware | Roadside

EH5 | East Herts Bishops Stortford | Roadside

EI0 | Ealing - Greenford | Urban Background

EI1 | Ealing - Western Avenue | Roadside

EL1 | Elmbridge - Bell Farm Hersham | Urban Background

EL2 | Elmbridge - Esher High Street | Roadside

EL3 | Elmbridge - Hampton Court Parade | Roadside

EL4 | Elmbridge - Walton High Street | Kerbside

EN1 | Enfield - Bushhill Park | Suburban

EN2 | Enfield - Church Street | Roadside

EN3 | Enfield - Salisbury School | Urban Background

EN4 | Enfield - Derby Road | Roadside

EN5 | Enfield - Bowes Primary School | Roadside

FB1 | Rushmoor - Medway Drive | Roadside

GB0 | Greenwich and Bexley - Falconwood FDMS | Roadside

GB6 | Greenwich and Bexley - Falconwood | Roadside

GL1 | Glasgow Centre | Suburban

GL4 | Glasgow Centre Aethalometer | Suburban

GN0 | Greenwich - A206 Burrage Grove | Roadside

GN2 | Greenwich - Millennium Village | Industrial

GN3 | Greenwich - Plumstead High Street | Roadside

GN4 | Greenwich - Fiveways Sidcup Rd A20 | Roadside

GR4 | Greenwich - Eltham | Suburban

GR5 | Greenwich - Trafalgar Road | Roadside

GR7 | Greenwich - Blackheath | Roadside

GR8 | Greenwich - Woolwich Flyover | Roadside

GR9 | Greenwich - Westhorne Avenue | Roadside

HA0 | Harwell - Carbon | Rural

HA1 | Harwell Rural AURN | Rural

HA2 | Harwell Rural PARTISOL | Rural

HA4 | Harwell Rural SMPS | Rural

HA9 | Harwell - Partisol SO4 | Urban Background

HF1 | Hammersmith and Fulham - Broadway | Roadside

HF2 | Hammersmith and Fulham - Brook Green | Urban Background

HF3 | Hammersmith and Fulham - Scrubs Lane | Kerbside

HG1 | Haringey - Haringey Town Hall | Roadside

HG2 | Haringey - Priory Park | Urban Background

HG3 | Haringey - Bounds Green | Roadside

HI0 | Hillingdon - Sipson Road | Suburban

HI1 | Hillingdon - South Ruislip | Roadside

HI2 | Hillingdon - Hillingdon Hospital | Roadside

HI3 | Hillingdon - Oxford Avenue | Roadside

HK4 | Hackney - Clapton | Urban Background

HK6 | Hackney - Old Street | Roadside

HL1 | Halifax Aethalometer | Urban Background

HM1 | Hertsmere Borehamwood 1 (Background) | Urban Background

HM4 | Hertsmere - Borehamwood | Urban Background

HO1 | Horsham Background | Urban Background

HO2 | Horsham - Park Way | Roadside

HO4 | Horsham - Storrington | Roadside

HO5 | Horsham - Cowfold | Roadside

HR1 | Harrow - Stanmore | Urban Background

HR2 | Harrow - Pinner Road | Roadside

HS1 | Hounslow - Brentford | Roadside

HS2 | Hounslow - Cranford | Suburban

HS3 | Hounslow - Brentford | Roadside

HS4 | Hounslow - Chiswick High Road | Roadside

HS5 | Hounslow - Brentford | Roadside

HS6 | Hounslow - Heston Road | Roadside

HS7 | Hounslow - Hatton Cross | Urban Background

HS9 | Hounslow - Feltham | Roadside

HT1 | Hastings - Bulverhythe | Roadside

HT2 | Hastings - Fresh Fields | Roadside

HV1 | Havering - Rainham | Roadside

HV2 | Havering - Harold Hill | Suburban

HV3 | Havering - Romford | Roadside

HX0 | Birmingham Tyburn Aethalometer | Urban Background

IC6 | City of London - Walbrook Wharf Indoor | Roadside

IG4 | Greenwich - Eltham Ecology Centre Indoor | Urban Background

IS1 | Islington - Upper Street | Urban Background

IS2 | Islington - Holloway Road | Roadside

IS4 | Islington - Foxham Gardens | Urban Background

IS5 | Islington - Duncan Terrace | Roadside

IS6 | Islington - Arsenal | Urban Background

IT2 | Tower Hamlets - Mile End Road | Roadside

KB1 | South Kirkby Aethalometer | Urban Background

KC0 | North Kensington - Carbon | Urban Background

KC1 | Kensington and Chelsea - North Ken | Urban Background

KC2 | Kensington and Chelsea - Cromwell Road | Roadside

KC3 | Kensington and Chelsea - Knightsbridge | Roadside

KC4 | Kensington and Chelsea - Kings Road | Roadside

KC5 | Kensington and Chelsea - Earls Court Rd | Kerbside

KC7 | Kensington and Chelsea - North Ken FDMS | Urban Background

KC9 | North Kensington - Partisol SO4 | Urban Background

KT1 | Kingston - Chessington | Suburban

KT2 | Kingston - Town Centre | Roadside

LA1 | Luton Airport | Urban Background

LB1 | Lambeth - Christchurch Road | Roadside

LB2 | Lambeth - Vauxhall Cross | Roadside

LB3 | Lambeth - Loughborough Junct | Urban Background

LB4 | Lambeth - Brixton Road | Kerbside

LB5 | Lambeth - Bondway Interchange | Roadside

LB6 | Lambeth - Streatham Green | Urban Background

LH0 | Hillingdon - Harlington | Urban Background

LH2 | Heathrow Airport | Urban Background

LL1 | Lullington Heath Rural AURN | Rural

LN1 | Luton - Challney Community College | Urban Background

LS1 | Lewes - Telscombe Cliffs | Roadside

LS2 | Lewes - Commercial Square | Roadside

LS4 | Newhaven - Denton School | Urban Background

LW1 | Lewisham - Catford | Urban Background

LW2 | Lewisham - New Cross | Roadside

LW3 | Lewisham - Mercury Way | Industrial

MA1 | Manchester Piccadilly CPC | Urban Background

MA2 | Manchester Piccadilly | Urban Background

MD1 | Mid Beds Biggleswade (Roadside) | Roadside

MD2 | Mid Beds Silsoe (Rural) | Rural

MD3 | Central Beds - Sandy | Roadside

MD4 | Central Beds - Marston Vale | Rural

ME1 | Merton - Morden Civic Centre | Roadside

MP1 | Marchwood Power - Marchwood | Industrial

MP2 | Marchwood Power - Millbrook Rd Soton | Industrial

MR3 | Marylebone Road Aethalometer | Kerbside

MV1 | Mole Valley - Leatherhead | Rural

MV2 | Mole Valley - Lower Ashtead | Suburban

MV3 | Mole Valley - Dorking | Urban Background

MW1 | Windsor and Maidenhead - Frascati Way | Roadside

MW2 | Windsor and Maidenhead - Clarence Road | Roadside

MW3 | Windsor and Maidenhead - Ascot | Rural

MY0 | Marylebone Road - Carbon | Kerbside

MY1 | Westminster - Marylebone Road | Kerbside

MY7 | Westminster - Marylebone Road FDMS | Kerbside

NA5 | Newtownabbey- Mallusk | Urban Background

NA6 | Newtownabbey- Shore Road | Roadside

NE2 | Port Talbot TEOM and CPC | Urban Background

NF1 | New Forest - Holbury | Industrial

NF2 | New Forest - Fawley | Industrial

NF3 | New Forest - Ringwood | Urban Background

NF4 | New Forest - Totton | Roadside

NF5 | New Forest - Lyndhurst | Roadside

NH1 | North Herts Mobile - Baldock 1 | Roadside

NH2 | North Herts Mobile - Baldock 2 | Roadside

NH3 | North Herts Mobile - Royston | Urban Background

NH4 | North Herts - Breechwood Green | Urban Background

NH5 | North Herts - Baldock Roadside | Roadside

NH6 | North Herts - Hitchin Library | Roadside

NK1 | North Kensington - CPC | Urban Background

NK3 | North Kensington Aethalometer | Urban Background

NK6 | North Kensington - URG | Urban Background

NM1 | Newham - Tant Avenue | Urban Background

NM2 | Newham - Cam Road | Roadside

NM3 | Newham - Wren Close | Urban Background

NW1 | Norwich Centre Aethalometer | Urban Background

OX0 | Oxford Centre Roadside AURN | Urban Background

OX1 | South Oxfordshire - Henley | Roadside

OX2 | South Oxfordshire - Wallingford | Roadside

OX3 | South Oxfordshire - Watlington | Roadside

OX4 | Oxford St Ebbes AURN | Urban Background

PO1 | Portsmouth Background AURN | Urban Background

PT6 | Port Talbot Dyffryn School | Industrial

RB1 | Redbridge - Perth Terrace | Urban Background

RB2 | Redbridge - Ilford Broadway | Kerbside

RB3 | Redbridge - Fullwell Cross | Kerbside

RB4 | Redbridge - Gardner Close | Roadside

RB5 | Redbridge - South Woodford | Roadside

RD0 | Reading AURN - New Town | Urban Background

RD1 | Reading - Caversham Road | Roadside

RD2 | Reading - Kings Road | Roadside

RD3 | Reading - Oxford Road | Roadside

RG1 | Reigate and Banstead - Horley | Suburban

RG2 | Reigate and Banstead - Horley South | Suburban

RG3 | Reigate and Banstead - Poles Lane | Rural

RG4 | Reigate and Banstead - Reigate High St | Kerbside

RHA | Richmond - Lower Mortlake Road | Roadside

RHB | Richmond - Lower Mortlake Road | Roadside

RI1 | Richmond - Castelnau | Roadside

RI2 | Richmond - Barnes Wetlands | Suburban

RI5 | Richmond Mobile - St Margarets | Kerbside

RI6 | Richmond Mobile - St Margarets | Kerbside

RI7 | Richmond Mobile - Richmond Park | Suburban

RI8 | Richmond Mobile - Richmond Park | Suburban

RIA | Richmond Mobile - George Street | Kerbside

RIB | Richmond Mobile - George Street | Kerbside

RIC | Richmond Mobile - Kew Rd | Kerbside

RID | Richmond Mobile - Kew Rd | Kerbside

RIE | Richmond Mobile - Richmond Rd Twickenham | Roadside

RIF | Richmond Mobile - Richmond Rd Twickenham | Roadside

RIG | Richmond Mobile - Upper Teddington Rd | Roadside

RIH | Richmond Mobile - Upper Teddington Rd | Roadside

RII | Richmond Mobile - Somerset Rd Teddington | Urban Background

RIJ | Richmond Mobile - Somerset Rd Teddington | Urban Background

RIK | Richmond Mobile - St. Margarets Grove | Urban Background

RIL | Richmond Mobile - St. Margarets Grove | Urban Background

RIM | Richmond Mobile - Petersham Rd Ham | Roadside

RIN | Richmond Mobile - Petersham Rd Ham | Roadside

RIO | Richmond Mobile - Stanley Rd Twickenham | Roadside

RIP | Richmond Mobile - Stanley Rd Twickenham | Roadside

RIQ | Richmond Mobile - Richmond Rd Twickenham | Roadside

RIR | Richmond Mobile - Richmond Rd Twickenham | Roadside

RIS | Richmond Mobile - Lincoln Ave Twickenham | Roadside

RIU | Richmond Mobile - Mortlake Rd Kew | Roadside

RIW | Richmond - Upper Teddington Road | Roadside

RIY | Richmond - Hampton Court Road | Kerbside

RO1 | Rochford - Rayleigh High Street | Roadside

RY1 | Rother - Rye Harbour | Rural

RY2 | Rother - De La Warr Road | Roadside

SA1 | St Albans - Fleetville | Urban Background

SB1 | South Beds - Dunstable | Urban Background

SC1 | Sevenoaks 1 | Suburban

SD1 | Southend-on-Sea AURN | Urban Background

SE1 | Stevenage - Lytton Way | Roadside

SH1 | Southampton Background AURN | Urban Background

SH2 | Southampton - Redbridge | Roadside

SH3 | Southampton - Onslow Road | Roadside

SH4 | Southampton - Bitterne | Urban Background

SK1 | Southwark - Larcom Street | Urban Background

SK2 | Southwark - Old Kent Road | Roadside

SK5 | Southwark - A2 Old Kent Road | Roadside

SL1 | Sunderland Aethalometer | Urban Background

ST1 | Sutton - Robin Hood School | Roadside

ST2 | Sutton - North Cheam | Urban Background

ST3 | Sutton - Carshalton | Suburban

ST4 | Sutton - Wallington | Kerbside

ST5 | Sutton - Beddington Lane | Industrial

ST6 | Sutton - Worcester Park | Kerbside

ST7 | Sutton - Therapia Lane | Industrial

SU1 | Sussex Mobile10 Stockbridge | Kerbside

SU2 | Sussex Mobile11 Jct Whitley Rd | Kerbside

SU3 | Sussex Mobile12 Cowfold | Kerbside

SU4 | Sussex Mobile 13 Newhaven | Roadside

SU5 | Sussex Mobile 14 Crawley | Roadside

SU6 | Sussex Mobile15 Chichester County Hall | Urban Background

SU7 | Sussex Mobile 16 Warnham | Rural

SU8 | Sussex Mobile 17 Newhaven Paradise Park | Roadside

SX1 | Sussex Mobile 1 | Urban Background

SX2 | Sussex Mobile 2 North Berstead | Roadside

SX3 | Sussex Mobile 3 | Roadside

SX4 | Sussex Mobile 4 Adur | Roadside

SX5 | Sussex Mobile 5 Fresh Fields Rd Hastings | Roadside

SX6 | Sussex Mobile 6 Orchard St Chichester | Roadside

SX7 | Sussex Mobile 7 New Road Newhaven | Roadside

SX8 | Sussex Mobile 8 Arundel | Kerbside

SX9 | Sussex Mobile 9 Newhaven Kerbside | Kerbside

TD0 | Richmond - National Physical Laboratory | Suburban

TE0 | Tendring St Osyth AURN | Rural

TE1 | Tendring - Town Hall | Roadside

TH1 | Tower Hamlets - Poplar | Urban Background

TH2 | Tower Hamlets - Mile End Road | Roadside

TH3 | Tower Hamlets - Bethnal Green | Urban Background

TH4 | Tower Hamlets - Blackwall | Roadside

TK1 | Thurrock - London Road (Grays) | Urban Background

TK2 | Thurrock - Purfleet | Roadside

TK3 | Thurrock - Stanford-le-Hope | Roadside

TK8 | Thurrock - London Road (Purfleet) | Roadside

TR1 | Three Rivers - Rickmansworth | Urban Background

UT1 | Uttlesford - Saffron Walden Fire Station | Roadside

UT2 | Uttlesford - Takeley | Urban Background

UT3 | Uttlesford - Broxted Farm | Rural

VS1 | Westminster - Victoria Street | Kerbside

WA1 | Wandsworth - Garratt Lane | Roadside

WA2 | Wandsworth - Town Hall | Urban Background

WA3 | Wandsworth - Roehampton | Rural

WA4 | Wandsworth - High Street | Roadside

WA6 | Wandsworth - Tooting | Roadside

WA7 | Wandsworth - Putney High Street | Kerbside

WA8 | Wandsworth - Putney High Street Facade | Roadside

WA9 | Wandsworth - Putney | Urban Background

WE0 | Kensington and Chelsea - Pembroke Road | Urban Background

WF1 | Watford (Roadside) | Roadside

WF2 | Watford - Watford Town Hall | Roadside

WH1 | Welwyn Hatfield - Council Offices | Urban Background

WL1 | Waltham Forest - Dawlish Road | Urban Background

WL2 | Waltham Forest - Mobile | Roadside

WL3 | Waltham Forest - Chingford | Roadside

WL4 | Waltham Forest - Crooked Billet | Kerbside

WL5 | Waltham Forest - Leyton | Roadside

WM0 | Westminster - Horseferry Road | Urban Background

WM3 | Westminster - Hyde Park Partisol | Roadside

WM4 | Westminster - Charing Cross Library | Roadside

WM5 | Westminster - Covent Garden | Urban Background

WM6 | Westminster - Oxford St | Kerbside

WR1 | Bradford Town Hall Aethalometer | Urban Background

WT1 | Worthing - Grove Lodge | Kerbside

XB1 | Bletchley | Rural

XS1 | Shukri Outdoor | Industrial

XS2 | Shukri Indoor | Industrial

XS3 | Osiris mobile | Urban Background

YH1 | Harrogate Roadside | Roadside

ZA1 | Ashford Rural - Pluckley | Rural

ZA2 | Ashford Roadside | Roadside

ZA3 | Ashford Background | Urban Background

ZA4 | Ashford M20 Background | Urban Background

ZC1 | Chatham Roadside - A2 | Roadside

ZD1 | Dover Roadside - Town Hall | Roadside

ZD2 | Dover Roadside - Townwall Street | Roadside

ZD3 | Dover Background - Langdon Cliff | Urban Background

ZD4 | Dover Background - East Cliff | Urban Background

ZD5 | Dover Coast Guard Met | Urban Background

ZD6 | Dover Docks | Industrial

ZF1 | Folkestone Suburban - Cheriton | Suburban

ZG1 | Gravesham Backgrnd - Northfleet | Urban Background

ZG2 | Gravesham Roadside - A2 | Roadside

ZG3 | Gravesham Ind Bgd - Northfleet | Urban Background

ZH1 | Thanet Rural - Minster | Rural

ZH2 | Thanet Background - Margate | Urban Background

ZH3 | Thanet Airport - Manston | Urban Background

ZH4 | Thanet Roadside - Ramsgate | Roadside

ZL1 | Luton Background | Urban Background

ZM1 | Maidstone Meteorological | Urban Background

ZM2 | Maidstone Roadside - Fairmeadow | Kerbside

ZM3 | Maidstone Rural - Detling | Rural

ZR1 | Dartford Roadside - St Clements | Kerbside

ZR2 | Dartford Roadside 2 - Town Centre | Roadside

ZR3 | Dartford Roadside 3 - Bean Interchange | Roadside

ZS1 | Stoke Rural AURN | Rural

ZT1 | Tonbridge Roadside - Town Centre | Roadside

ZT2 | Tunbridge Wells Background - Town Hall | Urban Background

ZT3 | Tunbridge Wells Rural - Southborough | Rural

ZT4 | Tunbridge Wells Roadside - St Johns | Roadside

ZT5 | Tonbridge Roadside 2 - High St | Roadside

ZV1 | Sevenoaks - Greatness Park | Urban Background

ZV2 | Sevenoaks - Bat and Ball | Roadside

ZW1 | Swale Roadside - Ospringe A2 | Roadside

ZW2 | Swale Background - Sheerness | Urban Background

ZW3 | Swale Roadside 2 - Ospringe Street | Roadside

ZY1 | Canterbury Backgrnd - Chaucer TS | Urban Background

ZY2 | Canterbury Roadside - St Dunstans | Roadside

ZY4 | Canterbury St Peters Place | Roadside

See also

Author

David Carslaw and Ben Barratt

Examples

## import all pollutants from Marylebone Rd from 1990:2009
if (FALSE) mary <- importKCL(site = "my1", year = 2000:2009)

## import nox, no2, o3 from Marylebone Road and North Kensignton for 2000
if (FALSE) thedata <- importKCL(site = c("my1", "kc1"), year = 2000,
pollutant = c("nox", "no2", "o3"))

## import met data too...
if (FALSE) my1 <- importKCL(site = "my1", year = 2008, met = TRUE)