Import monitoring site meta data for the AURN, KCL, SAQN, WAQN networks

Function to import meta data from UK air pollution monitoring sites

importMeta(source = "aurn", all = FALSE)

Arguments

source
The data source for the meta data. Can be "aurn", "kcl" or "saqn" (or "saqd"); upper or lower case.

all
When all = FALSE only the site code, site name, latitude and longitude and site type are imported. Setting all = TRUE will import all available meta data and provide details (when available) or the individual pollutants measured at each site.

Value

A data frame with meta data.

Details

This function imports site meta data from four networks in the UK: the Defra Automatic Urban and Rural Network (AURN), King's College London networks and the Scottish Air Quality Network. The meta data includes site location (latitude, longitude and OS easting and northing --- the latter for KCL networks), site type and it's start/close data, as well as other information.

The Scottish and Welsh air quality networks are available as SAQN (or SAQD) and WAQN.

The meta information can usefully be combined with matching air pollution data and produce maps of concentration --- see examples below. Note if many sites and/or years of hourly data are imported it may be better to aggregate first and then combine with the meta data information.

Thanks go to Dr Ben Barratt (KCL) and Trevor Davies (Riacrdo) for making these data available.

See also

Examples

## import AQ data and add meta data to data frame

# NOT RUN {
aq <- importAURN(site = c("kc1", "my1"), year = 2009)
meta <- importMeta(source = "aurn")
aq <- merge(aq, meta, by = "site")
# }

## aggregate first before adding meta data (useful for many sites/years)

# NOT RUN {
aq <- importAURN(site = c("kc1", "my1"), year = 2009)
meta <- importMeta(source = "aurn")
## calculate annual means
annual <- timeAverage(aq, avg.time = "year", type = "site")
annual <- merge(annual, meta, by = "site")
# }