Import monitoring site meta data for the UK and European networks

Function to import meta data for air quality monitoring sites

importMeta(source = "aurn", all = FALSE)

Arguments

source
The data source for the meta data. Can be “aurn”, “saqn” (or “saqd”), “aqe”, “ni”, “kcl” or “europe”; upper or lower case.

all
When all = FALSE only the site code, site name, latitude and longitude and site type are imported. Setting all = TRUE will import all available meta data and provide details (when available) or the individual pollutants measured at each site.

Value

A data frame with meta data.

Details

This function imports site meta data from several networks in the UK and Europe:

“aurn”, The UK Automatic Urban and Rural Network.

“saqn”, The Scottish Air Quality Network.

“waqn”, The Welsh Air Quality Network.

“ni”, The Northern Ireland Air Quality Network.

“aqe”, The Air Quality England Network.

“kcl”, King's College London networks.

“europe”, Import hourly European data (Airbase/e-reporting) based on a simplified version of the saqgetr package.

By default, the function will return the site latitude, longitude and site type. If the option all = TRUE is used, much more detailed information is returned. For most networks, this detailed information includes per-pollutant summaries, opening and closing dates of sites etc.

Thanks go to Trevor Davies (Ricardo), Dr Stuart Grange (EMPA) and Dr Ben Barratt (KCL) and for making these data available.

See also

Author

David Carslaw

Examples

## basic data

if (FALSE) {
meta <- importMeta(source = "aurn")

# more detailed information:
meta <- importMeta(source = "aurn", all = TRUE)

# from the Scottish Air Quality Network
meta <- importMeta(source = "saqn", all = TRUE)
}