Example data for openair

The mydata dataset is provided as an example dataset as part of the openair package. The dataset contains hourly measurements of air pollutant concentrations, wind speed and wind direction collected at the Marylebone (London) air quality monitoring supersite between 1st January 1998 and 23rd June 2005.

Format

Data frame with 65533 observations (rows) on the following 10 variables:

list("date")

Observation date/time stamp in year-month-day hour:minute:second format (POSIXct).

list("ws")

Wind speed, in m/s, as numeric vector.

list("wd")

Wind direction, in degrees from North, as a numeric vector.

list("nox")

Oxides of nitrogen concentration, in ppb, as a numeric vector.

list("no2")

Nitrogen dioxide concentration, in ppb, as a numeric vector.

list("o3")

Ozone concentration, in ppb, as a numeric vector.

list("pm10")

Particulate PM10 fraction measurement, in ug/m3 (raw TEOM), as a numeric vector.

list("so2")

Sulfur dioxide concentration, in ppb, as a numeric vector.

list("co")

Carbon monoxide concentration, in ppm, as a numeric vector.

list("pm25")

Particulate PM2.5 fraction measurement, in ug/m3, as a numeric vector.

Source

mydata was compiled from data archived in the London Air Quality Archive. See http://www.londonair.org.uk for site details.

The same data is also provide in '.csv' format via the openair project web site http://www.openair-project.org.

Details

mydata is supplied with the openair package as an example dataset for use with documented examples.

Note

openair functions generally require data frames with a field "date" that can be in either POSIXct or Date format but should be GMT time zone. This can be hourly data or higher resolution data.

Examples

#basic structure
head(mydata)

#>                  date   ws  wd nox no2 o3 pm10    so2      co pm25
#> 1 1998-01-01 00:00:00 0.60 280 285  39  1   29 4.7225  3.3725   NA
#> 2 1998-01-01 01:00:00 2.16 230  NA  NA NA   37     NA      NA   NA
#> 3 1998-01-01 02:00:00 2.76 190  NA  NA  3   34 6.8300  9.6025   NA
#> 4 1998-01-01 03:00:00 2.16 170 493  52  3   35 7.6625 10.2175   NA
#> 5 1998-01-01 04:00:00 2.40 180 468  78  2   34 8.0700  8.9125   NA
#> 6 1998-01-01 05:00:00 3.00 190 264  42  0   16 5.5050  3.0525   NA
#>