polar.annulus(openair)	R Documentation

Bivariate polar annulus plot

Description

Typically plots the concentration of a pollutant by wind direction and as a function of time as an annulus. The function is good for visualising how concentrations of pollutants vary by wind direction and a time period e.g. by month, day of week.

Usage


polar.annulus(polar,
pollutant = "nox",
resolution = "fine",
local.time = FALSE,
period = "hour",
type = "default",
limits = c(0, 100),
cols = "default",
width = "normal",
exclude.missing = TRUE,
date.pad = FALSE,
force.positive = TRUE,
k = 15,
key = TRUE,
main = "",
auto.text = TRUE, ...)

Arguments

polar 	 A data frame minimally containing ws, wd and a pollutant. Can also contain date if plots by time period are required.

pollutant 	 Mandatory. A pollutant or other variable name corresponding to a variable in a data frame should be supplied e.g. pollutant = "nox"

resolution 	 Two plot resolutions can be set: normal and fine (the default).

local.time 	 Should the results be calculated in local time? The default is TRUE. Emissions activity tends to occur at local time e.g. rush hour is at 8 am every day. When the clocks go forward in spring, the emissions are effectively released into the atmosphere at GMT - 1 hour during the summer. When plotting diurnal profiles, this has the effect of "smearing-out" the concentrations. A better approach is to express time as local time, which here is defined as BST (British Summer Time). This correction tends to produce better-defined diurnal profiles of concentration (or other variables) and allows a better comparison to be made with emissions/activity data. If set to FALSE then GMT is used.

period 	 This determines the temporal period to consider. Options are "trend" (the default), "season" to plot variation throughout the year, "weekday" to plot day of the week variation and "hour" to plot the diurnal variation.

type 	 The type can either be "default" or "site". The former will produce a single plot and the latter produces the same number of plots as there are sites. See the manual for details of use. It is also possible to choose type as another variable in the data frame. For example, type = "o3" will plot four wind roses for different levels of ozone, split into four quantiles (approximately equal numbers of counts in each of the four splits). This offers great flexibility for understanding the variation of different variables dependent on another.

limits 	 The functions does its best to choose sensible limits automatically. However, there are circumstances when the user will wish to set different ones. An example would be a series of plots showing each year of data separately. The limits are set in the form c(lower, upper), so limits = c(0, 100) would force the plot limits to span 0-100.

cols 	 Colours to be used for plotting. Options include "default", "increment", "heat", "jet" and user defined. For user defined the user can supply a list of colour names recognised by R (type colours() to see the full list). An example would be cols = c("yellow", "green", "blue")

width 	 The width of the annulus; can be "normal" (the default), "thin" or "fat".

exclude.missing 	 Setting this option to TRUE (the default) removes points from the plot that are too far from the original data. The smoothing routines will produce predictions at points where no data exist i.e. they predict. By removing the points too far from the original data produces a plot where it is clear where the original data lie. If set to FALSE missing data will be interpolated.

date.pad 	 For type = "trend" (default), date.pad = TRUE will pad-out missing data to the beginning of the first year and the end of the last year. The purpose is to ensure that the trend plot begins and ends at the beginning or end of year.

force.positive 	 The default is TRUE. Sometimes if smoothing data with streep gradients it is possible for predicted values to be negative. force.positive = TRUE ensures that predictions remain postive. This is useful for several reasons. First, with lots of missing data more interpolation is needed and this can result in artefacts because the predictions are too far from the original data. Second, if it is known beforehand that the data are all postive, then this option carries that assumption through to the prediction. The only likley time where setting force.positive = FALSE would be if background concentrations were first subtracted resulting in data that is legitimately negative. For the vast majority of situations it is expected that the user will not need to alter the default option.

k 	 This is the smoothing parameter that is set if auto.smooth is set to FALSE. Typically, value of around 15 (the default) seems to be suitable and will resolve more features in the plot. For type = "trend" k = 20 is used to reveal more detail. Increasing the value of k will decrease the amount of smoothing, the calculations will take longer. Increasing k for type other than trend will have little effect. However it can be useful in trends to show more detail, hence a higher default value of k. In some cases e.g. a trend plot with less than 1-year of data the smoothing with the default values may become too noisy and affected more by outliers. Choosing a lower value of k (say 10) may help produce a better plot.

key 	 Should a key be drawn?

main 	 Title of plot.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the ‘2’ in NO2.

... 	 Other graphical parameters.

Details

The polar.annulus function shares many of the properties of the polar.plot. However, polar.annulus is focussed on displaying information on how concentrations of a pollutant (values of another variable) vary with wind direction and time. Plotting as an annulus helps to reduce compression of information towards the centre of the plot. The circular plot is easy to interpret because wind direction is most easily understood in polar rather than Cartesian coordinates.

The inner part of the annulus represents the earliest time and the outer part of the annulus the latest time. The time dimension can be shown in many ways including "trend" (default), "hour" (hour or day), "season" (month of the year) and "weekday" (day of the week). Taking hour as an example, the plot will show how concentrations vary by hour of the day and wind direction. Such plots can be very useful for understanding how different source influences affect a location.

For type = "trend" the amount of smoothing does not vary linearly with the length of the time series i.e. a certain amount of smoothing per unit interval in time. This is a deliberate choice because should one be interested in a subset (in time) of data, more detail will be provided for the subset compared with the full data set. This allows users to investigate specific periods in morte detail. Full flexibility is given through the smoothing parameter k.

Warning

The function is written mostly for more than 1-year of hourly data. If less than 1-year of data are used and type = "trend" then type = "season" will be used, which should have the desired effect. Setting k too high may result in an error if there are insufficient data to justify such detailed smoothing. Calculations will take longer as k increases.

Author(s)

David Carslaw

See Also

polar.plot, polar.freq

Examples



# load example data from package
data(mydata)

# trend plot for PM10 at Marylebone Rd
polar.annulus(mydata, poll="pm10", main = "trend in pm10 at Marylebone Road")

# seasonal plot for PM10 at Marylebone Rd
## Not run: polar.annulus(mydata, poll="pm10", period = "season")

# trend in coarse particles (PMc = PM10 - PM2.5), calculate PMc first

mydata$pmc <- mydata$pm10 - mydata$pm25
## Not run: polar.annulus(mydata, poll="pmc", period = "trend", main = "trend in pmc at Marylebone Road")


[Package openair version 0.1 Index]