polar.percentile(openair)	R Documentation

Polar percentile plots

Description

Shows the concentration of a pollutant in polar coordinates and information the percentiles.

Usage


polar.percentile(polar,
pollutant = "",
type = "ws",
limits = c(0, 100),
per = c(50, 95),
cols = "default",
auto.smooth = TRUE,
k = 10,
main = "",
auto.text = TRUE, ...)

Arguments

polar 	 Mandatory. The user must supply a data frame with fields including date, pollutant, ws and wd.

pollutant 	 Mandatory. A pollutant name corresponding to a variable in a data frame should be supplied e.g. pollutant = "nox"

type 	 The type can be "ws" (default) "year" or "month", for plotting by wind speed ranges, year and by month of year. The choice of year is useful for examining trends and plotting by month can indicate seasonal influences.

limits 	 ~~Describe limits here~~

per 	 The user can provide overall percentile values to plot. This is is the form c(50, 95) e.g. to consider the 50th and 95th percentiles.

cols 	 Colours to be used for plotting. Options include "default", "increment", "heat", "spectral" and user defined.

auto.smooth 	 ~~Describe auto.smooth here~~

k 	 ~~Describe k here~~

main 	 ~~Describe main here~~

auto.text 	 ~~Describe auto.text here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

David Carslaw

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

[Package openair version 1.0 Index]